% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_eddy.R
\name{barplot_agg}
\alias{barplot_agg}
\title{Bar Plots of Aggregated Variables}
\usage{
barplot_agg(
  x,
  var,
  interval = NULL,
  nTicks = NULL,
  days = x$days,
  names.arg = x$Intervals
)
}
\arguments{
\item{x}{A data frame with column names.}

\item{var}{A character string. An \code{x} column name of the variable to
plot on y-axis.}

\item{interval}{A character string. Specifies \code{xlab} timescale and does
not affect computations (e.g. "daily", "3-daily", "weekly", "monthly",
etc.).}

\item{nTicks}{An integer. Number of x-axis ticks to plot. If \code{NULL},
maximum 20 ticks with corresponding \code{names.arg} will be plotted,
otherwise \code{nTicks} defaults to \code{8}. Specifying \code{nTicks}
allows greater control over the x-axis ticks density.}

\item{days}{A numeric vector. Number of days (or their fractions) aggregated
within each time interval described by \code{names.arg}. Used to specify
bar widths.}

\item{names.arg}{A character vector. Names of each aggregation period
corresponding to \code{x$var} used as x-axis labels.}
}
\description{
Creates a bar plot with aggregated variable on y-axis and labels of
aggregation periods on x-axis.
}
\examples{
set.seed(123)
n <- 17520 # number of half-hourly records in one non-leap year
tstamp <- seq(c(ISOdate(2021,3,20)), by = "30 mins", length.out = n)
x <- data.frame(timestamp = tstamp, H = rf(n, 1, 2, 1))
aggm <- agg_sum(x, "\%b-\%Y")
barplot_agg(aggm, var = "H_sum", "monthly")
aggw <- agg_sum(x, "\%W_\%Y")
barplot_agg(aggw, var = "H_sum", "weekly")
aggd <- agg_sum(x, "\%Y-\%m-\%d")
barplot_agg(aggd, var = "H_sum", "daily")

}
\seealso{
\code{\link{barplot}}
}
