% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quality_checking.R
\name{check_manually}
\alias{check_manually}
\title{Manual Data Flagging}
\usage{
check_manually(
  x,
  path,
  vars,
  qc_prefix,
  qc_suffix,
  interactive,
  siteyear,
  tname = "timestamp",
  shift.by = NULL,
  with_units = FALSE,
  win_size = 672,
  format = "\%Y-\%m-\%d \%H:\%M"
)
}
\arguments{
\item{x}{A data frame.}

\item{path}{A string. Specifies a path to directory where results should be
saved.}

\item{vars}{A character vector providing names of variables in \code{x} that
will be inspected.}

\item{qc_prefix, qc_suffix}{A string. Quality control columns corresponding to
\code{vars} names are required in \code{x} in format
\code{qc_prefix+vars+qc_suffix}.}

\item{interactive}{A logical value. If \code{TRUE}, manual checking will be
provided in an interactive session. If \code{FALSE}, previously created
file with manual flags will be reloaded or \code{NULL} will be returned.}

\item{siteyear}{A string. Unique label of inspected data set.}

\item{tname}{A string. Name of variable in \code{x} with date-time
information.}

\item{shift.by}{An integer value specifying the time shift (in seconds) to be
applied to the date-time information of the reloaded manual QC if present.}

\item{with_units}{A logical value indicating whether read (or written) data
frame with manual flags includes (should include) also units.}

\item{win_size}{An integer. Number of values displayed per plot.}

\item{format}{A string. Format of \code{tname} date-time information.}
}
\value{
A data frame with flags 0 (marking accepted points) and flags 2
  (marking excluded points). Results can be written to \code{path}.

  If \code{interactive = FALSE}, and no file in \code{path} with pattern
  \code{"manual_QC"}, \code{NULL} is returned.
}
\description{
A wrapper for \code{\link{exclude}} that allows visual inspection of selected
variables in a data frame and manual flagging of values to be discarded.
Saving and reloading of results is supported.
}
\details{
Automatic reload of previously saved results in \code{path} with filename
including pattern \code{"manual_QC"} is attempted. If found, timestamp is
merged with date-time information in \code{x} if not identical and quality
control is combined with the new flags marked by user (flag 2 marks data
exclusion). Proper alignment of timestamps can be assured by \code{shift.by}.
\code{path} is also used for saving file (\code{"manual_QC"} pattern) with
results. Actual flagging allows to run \code{\link{exclude}} over all
\code{vars}. Each variable is required to have associated quality control
column in format \code{qc_prefix+vars+qc_suffix}.

Function can be run in two modes. If \code{interactive = TRUE}, attempt to
load previously saved manual QC will be performed, user will be allowed to
flag data manually in interactive session and save (merged) results. If you
just want to reload previously saved results, use \code{interactive = FALSE}.
}
\examples{
\dontrun{
tstamp <- seq(c(ISOdate(2021,3,20)), by = "30 mins", length.out = 10)
x <- data.frame(timestamp = tstamp, qc_H = rep(0, 10), H = rnorm(10))
man <- check_manually(x, "./", vars = "H", qc_prefix = "qc_", qc_suffix = "",
interactive = TRUE, siteyear = "MySite", win_size = 10)
summary_QC(man, names(man)[-1])}

}
\seealso{
\code{\link{exclude}}, \code{\link{locator}},
  \code{\link{combn_QC}}, \code{\link{strptime_eddy}}, \code{\link{merge}}.
}
