% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_handling.R
\name{choose_avail}
\alias{choose_avail}
\title{Choose Available Names from Set}
\usage{
choose_avail(names, all_names)
}
\arguments{
\item{names}{A character vector with available names.}

\item{all_names}{A character vector with all expected variable names.}
}
\value{
A character vector with subset of expected variable names.
}
\description{
Available variable names are checked against the full set of expected
variables and missing cases are reported.
}
\examples{
all_names <- c("TA", "TS", "VPD", "LE", "H", "NEE")
names <- c("H", "LE", "PM10")
choose_avail(names, all_names)
}
