% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{coded_vars}
\alias{coded_vars}
\title{Coded Variables}
\format{
An object of class \code{data.frame} with 12 rows and 6 columns.
}
\usage{
coded_vars
}
\description{
Used within \code{\link{extract_QC}} to link names of filters (QC suffixes)
with EddyPro column names and relevant SA, GA, SAGA and ALL output names.
}
\keyword{internal}
