% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_handling.R
\name{combine_docu}
\alias{combine_docu}
\title{Combine Documentation}
\usage{
combine_docu(path)
}
\arguments{
\item{path}{A character vector. The full paths to TXT files.}
}
\description{
Read documentation from single or multiple TXT files. In case of multiple
files, combine them together with one additional line separating them.
}
\seealso{
\code{\link{readLines}}.
}
