% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_handling.R
\name{combn_QC}
\alias{combn_QC}
\title{Combine Quality Checking Results}
\usage{
combn_QC(x, qc_names, name_out, additive = FALSE, na.as = NA)
}
\arguments{
\item{x}{A data frame with column names.}

\item{qc_names}{A vector of names of data frame columns to combine.}

\item{name_out}{A character string providing \code{varnames} value of the
output.}

\item{additive}{A vector of logical values (\code{TRUE} or \code{FALSE}) for
the \code{qc_names} subset of \code{x} that determines if the flags should
be treated as additive (\code{additive = TRUE}) or with fixed effect
(\code{additive = FALSE}). If only one value is provided, all columns are
considered to be of the same type.}

\item{na.as}{A vector of numeric or \code{NA} values for the \code{qc_names}
subset of \code{x} that determines how should be the missing flags
interpreted. If only one value is provided, all columns are treated the
same way.}
}
\value{
A vector with attributes \code{varnames} and \code{units} is
  produced. \code{varnames} value is set by \code{name_out} argument and
  \code{units} value is set to default \code{"-"}.
}
\description{
Combine quality checking results depending whether they have a fixed or
cumulative effect or any combination of these effects. It is also checked how
should \code{NA}s be interpreted.
}
\details{
The quality checking results must be provided as a data frame with columns
containing quality flags resulting from individual tests/filters. For all
flags over all rows, the maximum and cumulative sum is taken for columns with
fixed and additive effect, respectively.

Typical values of argument \code{na.as} are \code{NA}, \code{0} or \code{2}.
\code{NA} value is only formal and does not suggest any change in
interpretation. Value \code{0} is used in the case that the \code{NA} value
of the quality test/filter is an expected result and means that the half-hour
was not checked by or has a good quality according to the given test/filter
(e.g. \code{\link{despikeLF}}). Value \code{2} would be used only if the user
wants to explicitly specify that \code{NA} flag value should not be used for
further analyses.
}
\examples{
set.seed(5)
aa <- data.frame(xx = sample(c(0:2, NA), 20, replace = T))
aa$yy <- sample(c(0:2, NA), 20, replace = T)
aa$add_F <- combn_QC(aa, qc_names = c("xx", "yy"), additive = F,
name_out = "add_F")
aa$add_T <- combn_QC(aa, qc_names = c("xx", "yy"), additive = T,
name_out = "add_T")
aa$add_F_na.as_0 <- combn_QC(aa, qc_names = c("xx", "yy"), additive = F,
na.as = 0, name_out = "add_F_na.as_0")
aa$add_F_na.as_0part <- combn_QC(aa, qc_names = c("xx", "yy"), additive = F,
na.as = c(0, NA), name_out = "add_F_na.as_0part")
aa$add_F_na.as_2 <- combn_QC(aa, qc_names = c("xx", "yy"), additive = F,
na.as = 2, name_out = "add_F_na.as_2")
}
\seealso{
\code{\link{summary_QC}}.
}
