% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_handling.R
\name{correct}
\alias{correct}
\title{Correct Character Vector}
\usage{
correct(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A corrected character vector.
}
\description{
Substitute given characters or strings by their alternatives.
}
\details{
Literal string \code{"(z-d)/L"} is renamed to \code{"zeta"}. Substitute
within strings: \itemize{\item \code{"co2_flux"} by \code{"NEE"} \item
\code{"*"} by \code{"star"} \item \code{"\%"} by \code{"perc"} \item
\code{"-"} and \code{"/"} by \code{"_"} \item round and square brackets by
empty string} using regular expression patterns.
}
\examples{
correct(c("[m]", "qc_co2_flux", "(z-d)/L", "x_70\%", "*[-(z-d)/L"))
}
