% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quality_checking.R
\name{desp_loop}
\alias{desp_loop}
\title{This function is not necessarily easy to use by hand. It is designed to work
well when called from higher level functions. End user should always use
zm().}
\usage{
desp_loop(SD_sub, date, nVals, z, c, plot = FALSE)
}
\description{
These are the basic computing engines called by lm used to fit linear models.
These should usually not be used directly unless by experienced users.
.lm.fit() is bare bone wrapper to the innermost QR-based C code, on which
glm.fit and lsfit are based as well, for even more experienced users.
}
