% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quality_checking.R
\name{despikeLF}
\alias{despikeLF}
\title{Low Frequency Data Despiking}
\usage{
despikeLF(x, flux, qc_flag, name_out, flux_thr = NULL, plot = FALSE,
  light = c("PAR", "GR"), night_thr = 0, nVals = 50, z = 7,
  c = 4.4478)
}
\arguments{
\item{x}{A data frame with column names representing required variables. See
'Details' below.}

\item{flux}{A character string. Specifies the column name in \code{x} with
flux values.}

\item{qc_flag}{A character string. Specifies the column name in \code{x} with
\code{flux} related quality control flag.}

\item{name_out}{A character string providing \code{varnames} attribute value
of the output.}

\item{flux_thr}{A numeric vector with 2 non-missing values. Specifies fixed
thresholds for \code{flux} column. Values outside this range will be
flagged as spikes (flag 2). If \code{flux_thr = NULL}, thresholds are not
applied.}

\item{plot}{A logical value. If \code{FALSE} (the default), a numeric vector
identifying spikes is produced. If \code{TRUE}, list of
\code{\link{ggplot}} objects visualizing the spikes is returned instead.}

\item{light}{A character string. Selects preferred variable for incoming
light intensity. \code{"PAR"} or \code{"GR"} is allowed. Can be
abbreviated. If \code{light = NULL}, \code{flux} values are not separated
to nighttime/daytime subsets and \code{night_thr} is not used.}

\item{night_thr}{A numeric value that defines the threshold  between night
(for \code{light} values equal or lower than \code{night_thr}) and day (for
\code{light} values higher than \code{night_thr}) for incoming light.}

\item{nVals}{A numeric value. Number of values withing 13 day blocks required
to obtain robust statistics.}

\item{z}{A numeric value. \eqn{MAD} scaling factor.}

\item{c}{A numeric value. \code{\link{mad}} \code{constant}.}
}
\value{
If \code{plot = FALSE}, an integer vector with attributes
  \code{"varnames"} and \code{"units"}. If \code{plot = TRUE}, a list of
  \code{ggplot} objects.
}
\description{
Scaled median absolute deviation from the median is applied to
double-differenced time series to identify outliers.
}
\details{
Low Frequency Data Despiking is not an additive QC test. \code{despikeLF}
follows the QC scheme using QC flag range 0 - 2. \code{varnames} attribute of
returned vector follows the 'Naming Strategy' described in
\code{\link{extract_QC}} and is distinguished by suffix spikesLF.

The data frame \code{x} is expected to have certain properties. It is
required that it contains column named \code{"timestamp"} of class
\code{"POSIXt"} with regular sequence of date-time values, typically with
(half-)hourly frequency. Any missing values in \code{"timestamp"} are not
allowed. Thus, if no records exist for given date-time value, it still has to
be included. It also has to contain required (depends on the argument values)
column names. If QC flags are not available for given flux, \code{qc_flag}
still has to be included in \code{x} as a named column with \code{0} values
(i.e. all values will be checked for outliers).

Only non-missing \code{flux} values with corresponing \code{qc_flag} values
below \code{2} are used to detect the outliers. Missing \code{flux} values or
those with assigned flag \code{2} or \code{NA} are not checked and marked by
\code{NA} flag in the output.

\code{flux_thr} is intended for exclusion of data clearly outside of
theoretically acceptable range for the whole dataset. If \code{flux_thr} is
specified, \code{flux} values below \code{flux_thr[1]} and above
\code{flux_thr[2]} are marked as spikes (flag 2) in the output. Such values
are further not used for computing statistics on double-differenced time
series.

\code{light} and \code{night_thr} are intended to separate data to nighttime
and daytime subsets with different statistical properties. Despiking is then
applied to individual subsets and combined QC flags are returned.

Despiking is done within blocks of 13 consecutive days to account for
seasonality of measured \code{flux}. Within each block, all records are
compared with its neighbours and \eqn{d[i]} scores are produced. This is
achieved by double-differencing: \deqn{d[i] = (flux[i] - flux[i-1]) -
(flux[i+1] - flux[i])} In order to obtain maximum amount of \eqn{d[i]}
scores, all missing flux values are removed from the block before \eqn{d[i]}
scores are produced. \code{flux} values are marked as spikes if \eqn{d[i]} is
higher (lower) than median of \eqn{d[i]} scores (\eqn{M[d]}) + (-) scaled
median absolute deviation: \deqn{d[i] > M[d] + (z * MAD / 0.6745)} \deqn{d[i]
< M[d] - (z * MAD / 0.6745)} MAD is defined as: \deqn{MAD = median(abs(d[i] -
M[d]))}

The algorithm tends to flag also values that are neighbours of spikes. To
prevent false flagging, \code{\link{median}} and \code{\link{mad}} of
\code{flux} values within given block (\eqn{M[flux]} and \eqn{mad[flux]},
respectively) is computed. Values can be marked as spikes only if
\deqn{flux[i] > M[flux] + (c * mad)} or \deqn{flux[i] < M[flux] - (c * mad)}

Number of available double-differenced flux values (\code{nVals}) is checked
within each block. If equal or below \code{nVals}, \eqn{d[i]} or
\eqn{flux[i]} values are checked against the statistics computed using entire
dataset to ensure robustness.
}
\section{Abbreviations}{
 \itemize{\item QC: Quality Control \item PAR:
  Photosynthetic Active Radiation [umol m-2 s-1] \item GR: Global Radiation
  [W m-2]}
}

\section{References}{
 Sachs, L., 1996. Angewandte Statistik: Anwendung
  Statistischer Methoden, Springer, Berlin.

  Papale, D., Reichstein, M., Canfora, E., Aubinet, M., Bernhofer, C.,
  Longdoz, B., Kutsch, W., Rambal, S., Valentini, R., Vesala, T., Yakir, D.,
  2006. Towards a more harmonized processing of eddy covariance CO2 fluxes:
  algorithms and uncertainty estimation. Biogeosciences Discuss. 3, 961-992.
  doi:10.5194/bgd-3-961-2006

  Mauder, M., Cuntz, M., Drue, C., Graf, A., Rebmann, C., Schmid, H.P.,
  Schmidt, M., Steinbrecher, R., 2013. A strategy for quality and uncertainty
  assessment of long-term eddy-covariance measurements. Agric. For. Meteorol.
  169, 122-135. doi:10.1016/j.agrformet.2012.09.006
}

\seealso{
\code{\link{combn_QC}}, \code{\link{extract_QC}},
  \code{\link{median}} and \code{\link{mad}}.
}
