% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eddy_data}
\alias{eddy_data}
\title{Eddy covariance data}
\format{
A data frame with 672 rows and 36 columns:
\describe{
\item{timestamp}{date-time information in POSIXct format}
\item{GR}{globar radiation}
\item{PAR}{photosynthetically active radiation}
\item{Rn}{net radiation}
\item{Tair}{air temperature}
\item{Tsoil}{soil temperature}
\item{RH}{relative humidity}
\item{VPD}{vapor pressure deficit}
\item{SWC}{soil water content}
\item{P}{precipitation}
\item{G}{soil heat flux}
\item{qc_G}{quality control information related to G}
\item{Tau}{momentum flux}
\item{qc_Tau_SSITC}{steady-state and integral turbulence characteristics
QC information related to Tau flux}
\item{qc_Tau_forGF}{QC information combined using all relevant QC filters
for Tau flux}
\item{H}{sensible heat flux}
\item{qc_H_SSITC}{steady-state and integral turbulence characteristics
QC information related to H flux}
\item{qc_H_forGF}{QC information combined using all relevant QC filters
for H flux}
\item{LE}{latent heat flux}
\item{qc_LE_SSITC}{steady-state and integral turbulence characteristics
QC information related to LE flux}
\item{qc_LE_forGF}{QC information combined using all relevant QC filters
for LE flux}
\item{NEE}{net ecosystem exchange}
\item{qc_NEE_SSITC}{steady-state and integral turbulence characteristics
QC information related to NEE flux}
\item{qc_NEE_forGF}{QC information combined using all relevant QC filters
for NEE flux}
\item{qc_NEE_forGF_UF}{QC information combined using all relevant QC
filters for NEE, including friction velocity (uStar) filtering}
\item{wind_speed}{wind speed}
\item{wind_dir}{wind direction}
\item{ustar}{friction velocity}
\item{H_f}{gap-filled H flux}
\item{LE_f}{gap-filled LE flux}
\item{NEE_uStar_f}{gap-filled NEE after application of uStar filtering}
\item{ET_f}{gap-filled evapotranspiration}
\item{Reco_uStar}{ecosystem respiration derived by nighttime-based
approach}
\item{GPP_uStar_f}{gross primary production derived by nighttime-based
approach}
\item{Reco_DT_uStar}{ecosystem respiration derived by daytime-based
approach}
\item{GPP_DT_uStar}{gross primary production derived by daytime-based
approach}
}
}
\usage{
eddy_data
}
\description{
A two week subset of eddy covariance and related meteorological data from the
mixed floodplain forest ecosystem station in Lanzhot, Czechia (CZ-Lnz).
}
\keyword{datasets}
