% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{essential_vars_QC}
\alias{essential_vars_QC}
\title{Quality Control Essential Variables}
\format{
An object of class \code{character} of length 56.
}
\usage{
essential_vars_QC
}
\description{
A set of variables useful when working with quality controlled data.
}
\keyword{datasets}
