% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_checking.R
\name{exclude}
\alias{exclude}
\title{Exclude Values}
\usage{
exclude(x, qc_x = NULL, y = NULL, z = NULL, name_out = "-", win_size = 672)
}
\arguments{
\item{x}{A numeric vector with values to inspect.}

\item{qc_x}{An integer vector in the range \code{0 - 2} providing quality
control information about \code{x}.}

\item{y, z}{A numeric vector with values of auxiliary variable helping with
the interpretation of \code{x}.}

\item{name_out}{A character string providing \code{varnames} attribute value
of the output.}

\item{win_size}{An integer. Number of \code{x} values displayed per plot.}
}
\value{
An integer vector with attributes \code{"varnames"} and
\code{"units"}.
}
\description{
Interactive plots are used for identification and flagging of data for
exclusion based on the visual inspection.
}
\details{
Six options are available during the interactive session.
\enumerate{
\item flag values: select interval of values to exclude or use double-click to
flag single value.
\item undo last: allows to revert previous flagging (only one step back).
\item refresh plots: apply current user
flagging input to the plots. It removes excluded points and affects y-axis
range or it can show again excluded points if option 2 (undo last) was
applied before.
\item next plot.
\item jump to plot: any plot within the
existing range (see as a plot title) can be selected.
\item finalize: finish the flagging and return the results.
}

The manual flagging is finished only after it is \emph{finalized} (option 6).
The interactive session will be finished successfully also when
option 4 (\emph{next plot}) is executed while at the last plot. If \code{exclude()}
is \code{Escape}d before it is finalized, the user will lose the whole progress
(see \emph{undo last} option instead).
}
\examples{
a <- eddy_data[c("NEE", "qc_NEE_SSITC", "GR", "Tair")]

# include spikes
a$NEE[c(152, 479)] <- c(231, -2000)

# flag data manually (no auxiliary data)
if (dev.interactive()) {
  exclude(x = a$NEE, qc_x = a$qc_NEE_SSITC)
}

# flag data manually (single auxiliary variable)
if (dev.interactive()) {
  exclude(x = a$NEE, qc_x = a$qc_NEE_SSITC, y = a$GR)
}

# flag data manually (two auxiliary variables)
if (dev.interactive()) {
  exclude(x = a$NEE, qc_x = a$qc_NEE_SSITC, y = a$GR, z = a$Tair)
}

}
\seealso{
\code{\link[=check_manually]{check_manually()}}
}
