% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quality_checking.R
\name{extract_coded}
\alias{extract_coded}
\title{Extract Quality Control Information from Coded Values}
\usage{
extract_coded(x, prefix = "[8]", split = "[/]")
}
\arguments{
\item{x}{An atomic type.}

\item{prefix}{Character string containing a \code{\link{regular expression}}
identifying the prefix of coded values.}

\item{split}{Character string containing a \code{\link{regular expression}}
identifying the separator of variable names in \code{units} attribute.}
}
\value{
A data frame with columns \code{"SA"} and \code{"SAGA"}. Each
  column has attributes \code{"varnames"} and \code{"units"} and length equal
  to that of \code{x}.
}
\description{
This function is called by \code{\link{extract_QC}} and is not inteded to be
used directly. QC information for multiple variables stored in a character
vector of certain properties is extracted and interpreted.
}
\details{
Each element of \code{x} is expected to have the same structure according to
the format specified by \code{units} attribute of \code{x}. \code{units}
format string (e.g. \code{"8u/v/w/ts/h2o/co2"}) starts with a prefix
(\code{"8"}) that is followed by variable names (\code{"u", "v", "w", "ts",
"h2o", "co2"}) that are distinguished by a separator (\code{"/"}). Elements
of \code{x} thus consist of either seven components (prefix and six flags for
each variable) or one \code{NA} value. Flags can have three values: \itemize{
\item Flag 0: measured variable passed the test. \item Flag 1: measured
variable failed the test. \item Flag 9: flag could not be obtained for the
given variable (\code{NA}).}

\code{x} is interpreted in respect to instruments required to measure given
fluxes. SA is required for measurements of all fluxes and solely provides
data for computation of Tau and H fluxes. A combination of SA and GA (SAGA)
is needed for measurements of LE and NEE. To confirm the correct performance
of SA, all variables \code{"u", "v", "w", "ts"} must have flag 0. In case of
SAGA, all variables \code{"u", "v", "w", "ts", "h2o", "co2"} must have flag
0. Results are reported according to the QC scheme using QC flag range 0 - 2.
}
\section{Abbreviations}{
 \itemize{ \item QC: Quality Control \item SA: Sonic
  Anemometer \item GA: Gas Analyzer \item Tau: Momentum flux [kg m-1 s-2]
  \item H: Sensible heat flux [W m-2] \item LE: Latent heat flux [W m-2]
  \item NEE: Net ecosystem exchange [umol m-2 s-1] \item u: Longitudinal wind
  speed component [m s-1] \item v: Cross-wind wind speed component [m s-1]
  \item w: Vertical wind speed component [m s-1] \item ts: Sonic temperature
  [degC] \item h2o: H2O concentration [mmol mol-1] \item co2: CO2
  concentration [umol mol-1]}.
}

\examples{
set.seed(5)
xx <- replicate(10,
                paste0(c("8", sample(c(0:1, 9), 6, c(0.7, 0.2, 0.1),
                                    replace = TRUE)),
                      collapse = ""))
is.na(xx) <- c(2, 5)
units(xx) <- "8u/v/w/ts/h2o/co2"
cbind(xx, extract_coded(xx))
}
\seealso{
\code{\link{extract_QC}}.
}
