% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{fluxes}
\alias{fluxes}
\title{Supported Fluxes}
\format{
An object of class \code{character} of length 4.
}
\usage{
fluxes
}
\description{
A complete set of supported fluxes in eddy covariance workflow.
}
\details{
While users can utilize \code{\link[openeddy]{openeddy}} to create their own
processing workflow for any set of fluxes or variables, the original intent
was to process specified fluxes. The processing workflow is available here:
\url{https://github.com/lsigut/EC_workflow}.
}
\keyword{datasets}
