% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_handling.R
\name{get_varnames}
\alias{get_varnames}
\alias{get_units}
\alias{set_varnames}
\alias{set_units}
\title{Get or Set Attributes Varnames and Units}
\usage{
get_varnames(x, names = FALSE)

get_units(x, names = FALSE)

set_varnames(x, varnames)

set_units(x, units)
}
\arguments{
\item{x}{A data frame or an atomic type.}

\item{names}{A logical value. Applies only in case of data frames. If
\code{TRUE}, attributes are extracted with corresponding column names.}

\item{varnames, units}{An atomic type. The length must be \code{1} if \code{x}
is an atomic type or equal to number of columns in \code{x} if \code{x} is
a data frame.}
}
\value{
For \code{get_varnames} and \code{get_units}, a character string or
  vector.

  For \code{set_varnames} and \code{set_units}, object \code{x} with set or
  updated attributes.
}
\description{
\code{varnames} and \code{units} are useful attributes that can store
original variable names (\code{varnames}) and units of measurement
(\code{units}) of each column in a data frame or of an atomic type. These
attributes can be displayed or set by following functions.
}
\details{
Functions check whether the extracted or set attributes contain elements with
\code{NULL}, \code{NA}, \code{""} values or if length of each element is
higher than \code{1}. In these cases, such elements are substituted with
\code{"-"}.
}
\examples{
xx <- data.frame(a = 1, b = 2, c = 3, d = 4)
lapply(xx, attr, "units")
get_units(xx, names = TRUE)
xx <- set_varnames(xx, c("a", "", NA, "d"))
xx <- set_units(xx, 1:4)
str(xx)
units <- get_units(xx)

## NB: subsetting by rows removes 'varnames' and 'units' attributes
str(yy <- xx[1, ])
yy <- set_varnames(yy, names(yy))
yy <- set_units(yy, units)
str(yy)
}
\seealso{
\code{\link{read_eddy}} and \code{\link{write_eddy}}.
}
