% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summarizing.R
\name{Griebel20_budgets}
\alias{Griebel20_budgets}
\title{Compute Griebel et al. (2020) budgets}
\usage{
Griebel20_budgets(
  df,
  TimestampCol = "TIMESTAMP_START",
  wdCol = "WD",
  targetCol = "NEE_VUT_REF",
  QcCol = "NEE_VUT_REF_QC",
  interval = 1800L,
  flux = c("carbon", "energy"),
  nInt = 8L,
  year = NULL,
  NEE_scor = TRUE,
  GPP_scor = FALSE,
  normalize = TRUE
)
}
\arguments{
\item{df}{A data frame.}

\item{TimestampCol}{A character string. Specifies a column name in \code{df}
that carries date-time information either in \code{POSIXt} or text strings
of format \code{"\%Y\%m\%d\%H\%M"}. Date-time information is expected to
represent either start or middle of the averaging period.}

\item{wdCol}{A character string. Specifies a column name in \code{df} that
carries the wind direction in degrees.}

\item{targetCol}{A character string. Specifies a column name in \code{df}
that carries the flux values for budget computations.}

\item{QcCol}{A character string or \code{NULL}. Specifies a column name in
\code{df} that carries gap-filling quality flags of \code{targetCol}
variable. It is assumed that \code{df[, QcCol] == 0} identifies the
measured (not gap-filled) records of \code{targetCol} variable. If
\code{NULL}, all non-missing values of \code{targetCol} are used for
budgeting.}

\item{interval}{An integer value. Represents an extent of eddy covariance
averaging period in seconds (e.g. 1800 for 30 mins, 3600 for 60 mins).}

\item{flux}{A character string. What type of flux does \code{targetCol}
represent? Can be abbreviated.}

\item{nInt}{An integer value. A number of wind sectors and time intervals for
binning.}

\item{year}{An integer vector. If \code{NULL}, budgets are produced for all
years available in \code{df}. Otherwise only specified years are processed.}

\item{NEE_scor, GPP_scor}{A logical value. Should sign correction of NEE (GPP)
be performed?}

\item{normalize}{A logical value. If \code{TRUE} (default), space and
space-time equitable budgets are corrected for the missing number of
records in a year.}
}
\value{
A data frame with columns corresponding to year, different types of
  budgets and number of observations used for budget computation each year.
  Each column has assigned attributes varnames and units.
}
\description{
Yearly budgets with different consideration of space-time equity.
}
\details{
The function produces several variants of budgets that represent annual sums
of measured and quality checked flux with different consideration of
space-time equity. In order to obtain budgets in sensible units after
summation, appropriate \code{flux} type must be specified. E.g. conversion
factor for carbon fluxes (umol(CO2) m-2 s-1 -> g(C) m-2 s-1) is 12.0107e-6,
conversion factor for energy fluxes (W m-2 -> MJ m-2 s-1) is 1e-6. Temporal
aspect of the conversion is handled based on \code{interval} extent.

Available variants of budgets include Traditional budget (uncorrected sum of
measured fluxes), Standardized budget (corrected according to wind sector
climatology based on all observation years), Space-equitable budget (each
sector contributes the exact same amount to budget) and Space-time-equitable
budget (each sector contributes equally to budget and sector contributions
are made time-uniform). Computation is generalized for any number of
\code{nInt} and any extent of \code{interval}. Please notice that Traditional
budget and Standardized budget differ only if multiple years are used for
computation. The reliability of the results depends on the data availability
within each year. For details see \code{References}.

Arguments specifying \code{df} column names represent FLUXNET standard. To
process \code{REddyProc} outputs, timestamp must be corrected to represent
middle of averaging period and appropriate columns selected (see
\code{Examples}).
}
\section{Sign Correction}{
 Although common sign convention for measured NEE
  (Net Ecosystem Exchange) denotes negative fluxes as CO2 uptake, summed NEE
  is typically reported with the opposite sign convention and is assumed to
  converge to NEP (Net Ecosystem Production), especially over longer
  aggregation intervals. In case of GPP (Gross Primary Production),
  \code{REddyProc} package applies sign convention denoting positive fluxes
  as carbon sink, thus sign correction before summing is not needed.

  Since there is no reliable way to guess the sign convention used in the
  data set, \code{NEE_scor} and \code{GPP_scor} must be specified. The
  default values (\code{NEE_scor = TRUE}; \code{GPP_scor = FALSE}) are
  adapted to sign convention applied in \code{REddyProc} package.
}

\section{References}{
  Griebel, A., Metzen, D., Pendall, E., Burba, G., &
  Metzger, S. (2020). Generating spatially robust carbon budgets from flux
  tower observations. Geophysical Research Letters, 47, e2019GL085942.
  https://doi.org/10.1029/2019GL085942
}

\examples{
\dontrun{
library(REddyProc)

# convert timestamp
DETha98 <- fConvertTimeToPosix(Example_DETha98, 'YDH', Year = 'Year',
Day = 'DoY', Hour = 'Hour')[-(2:4)]

# generate randomly wind directions for demonstration purposes (not included)
DETha98$WD <- sample(0:359, nrow(DETha98), replace = TRUE)

# if QcCol = NULL, all non-missing values of targetCol are used for budgeting
not_filled <- DETha98
not_filled$DateTime <- not_filled$DateTime - 900
Griebel20_budgets(not_filled, "DateTime", "WD", "LE", NULL, flux = "energy")

# gap-filling is not needed but illustrates processing of FLUXNET data
# notice that ustar filtering of NEE should be applied before budgeting
DETha98 <- filterLongRuns(DETha98, "NEE")
EProc <- sEddyProc$new('DE-Tha', DETha98,
c('NEE', 'Rg', 'Tair', 'VPD', 'Ustar'))
EProc$sMDSGapFillAfterUstar('NEE', uStarTh = 0.3, FillAll = TRUE)
filled <- cbind(DETha98, EProc$sExportResults())

# correct timestamp to represent middle of averaging period
filled$DateTime <- filled$DateTime - 900
Griebel20_budgets(filled, "DateTime", "WD", "NEE", "NEE_uStar_fqc")
}

}
\seealso{
\code{\link{spti_boot}} and \code{\link{spti_coverage}}.
}
