% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_checking.R
\name{interdep}
\alias{interdep}
\title{Flux Interdependency}
\usage{
interdep(qc_LE, qc_H = NULL, IRGA = c("en_closed", "open"), used_IRGA = NULL)
}
\arguments{
\item{qc_LE}{An integer vector. Latent heat quality control flags combining
all filters relevant for detection of instrument failure.}

\item{qc_H}{An integer vector. Sensible heat quality control flags combining
all filters relevant for detection of instrument failure. Used only if
\code{IRGA = "open"}.}

\item{IRGA}{A character string. Specifies the type of IRGA. Allowed values
are "en_closed" both for closed and enclosed path systems and
"open" for open path systems. Can be abbreviated.}

\item{used_IRGA}{A character vector or \code{NULL}. If specified, user-provided
IRGA type (either "en_closed" or "open"; cannot be abbreviated) used within
each averaging period is expected if IRGA type changed throughout the data
set. Any other \code{used_IRGA} value will be considered missing and assigned
flag 0 (relaxed approach). \code{IRGA} is ignored if \code{used_IRGA} is specified.}
}
\value{
A data frame. Each column has attributes \code{"varnames"} and
\code{"units"}.
}
\description{
Interdependency of H, LE and NEE QC flags due to corrections/conversions.
}
\details{
Flux interdependency is an additive QC flag correction. Results follow the QC
scheme using QC flag range 0 - 2. Returned data frame follows the 'Naming
Strategy' described in \code{\link[=extract_QC]{extract_QC()}}. Returned QC flags have QC
suffix interdep.

To convert buoyancy flux to sensible heat flux (SND or Schotanus correction),
reliable measurements of LE must be available. To correct LE and NEE
estimated by open-path IRGA (\code{IRGA = "open"}) for the effects of density
fluctuations due to temperature and humidity fluctuations (WPL or Webb
correction), reliable measurements of H must be available. To perform WPL
correction of NEE estimated with any kind of IRGA, reliable measurements of
LE must be available. Thus following set of rules apply:

If \code{IRGA = "en_closed"}
\itemize{
\item If \code{qc_LE == 2 | is.na(qc_LE) == TRUE}: qc_H and qc_NEE flags are increased
by 1.
}

If \code{IRGA = "open"}
\itemize{
\item if \code{qc_LE == 2 | is.na(qc_LE) == TRUE}: qc_H flags are increased by 1.
\item If \code{qc_H == 2 | is.na(qc_H) == TRUE}: qc_LE flags are increased by 1.
\item If \code{qc_H == 2 | is.na(qc_H) == TRUE | qc_LE == 2 | is.na(qc_LE) == TRUE}:
qc_NEE flags are increased by 1.
}

Input is not provided via a data frame on purpose. Required QC flags
typically need to be computed and are not part of the original EddyPro file.
These QC flags also serve no other purpose than for \code{interdep()} and are
not further used.
}
\section{Abbreviations}{

\itemize{
\item QC: Quality Control
\item H: Sensible heat flux [W m-2]
\item LE: Latent heat flux [W m-2]
\item NEE: Net ecosystem exchange [umol m-2 s-1]
\item IRGA: Infrared Gas Analyzer.
}
}

\section{References}{
 Mauder, M., Cuntz, M., Drue, C., Graf, A., Rebmann, C.,
Schmid, H.P., Schmidt, M., Steinbrecher, R., 2013. A strategy for quality
and uncertainty assessment of long-term eddy-covariance measurements.
Agric. For. Meteorol. 169, 122-135.
\url{https://doi.org/10.1016/j.agrformet.2012.09.006}
}

\examples{
interdep(0:2, IRGA = "en_closed")
interdep(0:2, 2:0, IRGA = "open")
interdep(0:2, 2:0, used_IRGA = c("enclosed", "open", "en_closed"))

}
\seealso{
\code{\link[=label_periods]{label_periods()}}, \code{\link[=combn_QC]{combn_QC()}} and \code{\link[=extract_QC]{extract_QC()}}.
}
