% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_checking.R
\name{label_periods}
\alias{label_periods}
\title{Label Periods}
\usage{
label_periods(x, labels, start, end = NULL, name_out = "-")
}
\arguments{
\item{x}{A POSIXt vector providing timestamp of measurements.}

\item{labels}{An \link{atomic} type.}

\item{start, end}{A POSIXt vector marking the start (end) of the periods to
be labelled. If \code{end} is specified, they must be of the same length.}

\item{name_out}{A character string providing \code{varnames} attribute value
of the output.}
}
\value{
An atomic type with the same length as \code{x}. Its
\code{varnames} and \code{units} attributes are set to  \code{name_out} and
\code{"-"} values, respectively.
}
\description{
Annotate selected periods with custom labels.
}
\details{
Labels are always only in the extent of \code{x}, although the range of labeled
period provided by \code{start} and \code{end} can surpass it.

If \code{end = NULL} (default), \code{end} is inferred from the \code{start} argument. In
that case \code{start} provides the break points between the labelled periods and
the last period extends to the last timestamp in \code{x}. If \code{end} is specified,
gaps between periods with no labels can exist.
}
\examples{
# alternative style: as.POSIXct("2000-01-01 12:15:00", tz = "GMT")
timestamp <- seq(ISOdatetime(2000, 1, 1, 12, 15, 0, tz = "GMT"),
                 ISOdatetime(2000, 1, 1, 18, 15, 0, tz = "GMT"),
                 by = "30 mins")

start <- c(ISOdatetime(2000, 1, 1, 12, 15, 0, tz = "GMT"),
           ISOdatetime(2000, 1, 1, 15, 15, 0, tz = "GMT"))

# specify labels that can be useful for extract_QC() or interdep()
rec <- label_periods(timestamp, c(18000, 36000), start,
                     name_out = "max_records")
rot <- label_periods(timestamp, c("double", "planar fit"), start,
                     name_out = "used_rotation")
IRGA <- label_periods(timestamp, c("open", "en_closed"), start,
                      name_out = "used_IRGA")
data.frame(timestamp, max_records = rec, used_rotation = rot,
           used_IRGA = IRGA)

}
\seealso{
\code{\link[=extract_QC]{extract_QC()}}, \code{\link[=interdep]{interdep()}}.
}
