% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_checking.R
\name{mf}
\alias{mf}
\title{Missing fraction of records}
\usage{
mf(x, ur, mfr)
}
\arguments{
\item{x}{A data frame. Contains two columns with the count of high frequency
spikes detected for the pair of variables used to compute covariance.}

\item{ur}{An integer vector. Number of used records for computation of
covariance.}

\item{mfr}{An integer value. Maximum number of records in a file (per whole
dataset).}
}
\description{
Computes missing fraction of high frequency records for particular flux.
}
\keyword{internal}
