% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openeddy.R
\docType{package}
\name{openeddy}
\alias{openeddy-package}
\alias{openeddy}
\title{Post-process eddy covariance data with ease}
\description{
The openeddy package is a software infrastructure for eddy covariance data
post-processing. It provides tools for handling data with attached metadata,
setting up quality control scheme and summarizing and plotting the results.
It is aligned with REddyProc package that provides methods for
uStar-filtering, gap-filling, and flux-partitioning. Thus the combined use of
openeddy and REddyProc allows to run the whole eddy covariance
post-processing chain. Learn more at <https://github.com/lsigut/EC_workflow>.
}
\section{References}{
 Mauder, M., Cuntz, M., Drue, C., Graf, A., Rebmann, C.,
  Schmid, H.P., Schmidt, M., Steinbrecher, R., 2013. A strategy for quality
  and uncertainty assessment of long-term eddy-covariance measurements.
  Agric. For. Meteorol. 169, 122-135. doi:10.1016/j.agrformet.2012.09.006
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lsigut/openeddy}
  \item Report bugs at \url{https://github.com/lsigut/openeddy/issues}
}

}
\author{
\strong{Maintainer}: Ladislav Sigut \email{sigut.l@czechglobe.cz}

Other contributors:
\itemize{
  \item Matthias Mauder \email{matthias.mauder@kit.edu} [contributor]
  \item Milan Fischer \email{fischer.milan@gmail.com} [contributor]
  \item Thomas Wutzler \email{twutz@bgc-jena.mpg.de} [contributor]
  \item Pavel Sedlak \email{sedlak@ufa.cas.cz} [contributor]
  \item Lenka Mzourkova Macalkova \email{mzourkova.l@gmail.com} [contributor]
  \item Radek Czerny \email{czerny.r@czechglobe.cz} [contributor]
  \item Ryan Patrick McGloin \email{mcgloin.r@czechglobe.cz} [contributor]
  \item Jiri Dusek \email{dusek.j@czechglobe.cz} [contributor]
  \item Natalia Kowalska \email{kowalska.n@czechglobe.cz} [contributor]
  \item Georg Jocher \email{jocher.g@czechglobe.cz} [contributor]
  \item Lenka Foltynova \email{foltynova.l@czechglobe.cz} [contributor]
  \item Marian Pavelka \email{pavelka.m@czechglobe.cz} [contributor]
  \item Gerardo Fratini \email{Gerardo.Fratini@licor.com} [contributor]
  \item Global Change Research Institute CAS [copyright holder]
  \item University of Ostrava [copyright holder]
}

}
