% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{precheck_vars}
\alias{precheck_vars}
\title{Precheck Variables}
\format{
An object of class \code{character} of length 46.
}
\usage{
precheck_vars
}
\description{
A set of variables typically available in EddyPro full output that can be
useful for preliminary check before quality control procedure.
}
\keyword{datasets}
