% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quality_checking.R
\name{separate}
\alias{separate}
\title{Helper function utilized in extract_coded() for abslim and spikesHF
- controls column splitting}
\usage{
separate(x, ...)
}
\description{
Helper function utilized in extract_coded() for abslim and spikesHF
- controls column splitting
}
\keyword{internal}
