% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summarizing.R
\name{spti_boot}
\alias{spti_boot}
\title{Space-time-equitable budgets with bootstrapping}
\usage{
spti_boot(
  df,
  TimestampCol = "TIMESTAMP_START",
  wdCol = "WD",
  targetCol = "NEE_VUT_REF",
  QcCol = "NEE_VUT_REF_QC",
  interval = 1800L,
  conv_fac = 1.20107e-05,
  nInt = 8L,
  year = NULL,
  samples = 100L,
  normalize = TRUE
)
}
\arguments{
\item{df}{A data frame.}

\item{TimestampCol}{A character string. Specifies a column name in \code{df}
that carries date-time information either in \code{POSIXt} or text strings
of format \code{"\%Y\%m\%d\%H\%M"}. Date-time information is expected to
represent either start or middle of the averaging period.}

\item{wdCol}{A character string. Specifies a column name in \code{df} that
carries the wind direction in degrees.}

\item{targetCol}{A character string. Specifies a column name in \code{df}
that carries the flux values for budget computations.}

\item{QcCol}{A character string or \code{NULL}. Specifies a column name in
\code{df} that carries gap-filling quality flags of \code{targetCol}
variable. It is assumed that \code{df[, QcCol] == 0} identifies the
measured (not gap-filled) records of \code{targetCol} variable. If
\code{NULL}, all non-missing values of \code{targetCol} are used for
budgeting.}

\item{interval}{An integer value. Represents an extent of eddy covariance
averaging period in seconds (e.g. 1800 for 30 mins, 3600 for 60 mins).}

\item{conv_fac}{A numeric value. Unit conversion factor for \code{targetCol}
variable that allows to produce budgets with meaningful units after
summation. Temporal aspect of conversion is handled internally based on
\code{interval} extent. E.g. for carbon fluxes (umol(CO2) m-2 s-1 -> gC m-2
s-1) \code{conv_fac = 12.0107e-6}, for energy fluxes (W m-2 -> MJ m-2 s-1)
\code{conv_fac = 1e-6}.}

\item{nInt}{An integer value. A number of wind sectors and time intervals for
binning.}

\item{year}{An integer vector. If \code{NULL}, budgets are produced for all
years available in \code{df}. Otherwise only specified years are processed.}

\item{samples}{An integer value. Amount of bootstraps to produce.}

\item{normalize}{A logical value. If \code{TRUE} (default), space and
space-time equitable budgets are corrected for the missing number of
records in a year.}
}
\value{
A data frame with columns corresponding to year, space-time-equitable
  budget of original dataset (space_time_eq_orig), 5\%, 50\% and 95\%
  quantiles of resampled datasets (space_time_eq_q05, space_time_eq_q50,
  space_time_eq_q95) and number of observations used for budget computation
  each year.
}
\description{
Yearly space-time-equitable budgets with uncertainty estimation.
}
\details{
Data from individual years are separated to \code{nInt} number of bins (e.g.
for \code{nInt = 8} that is \code{8x8 = 64} bins) as in the original Griebel
et al. (2020) method. Each bin is then resampled \code{samples} amount of
times with \code{\link{sample}} \code{size} according to the smallest amount
of records across all bins. In addition to space-time-equitable budget of
original dataset (space_time_eq_orig), 5\%, 50\% and 95\% quantiles of
resampled datasets (space_time_eq_q05, space_time_eq_q50, space_time_eq_q95)
are provided for uncertainty assessment.

Space-time-equitable budgeting assures that each sector contributes equally
to budget and sector contributions are made time-uniform. Computation is
generalized for any number of \code{nInt} and any extent of \code{interval}.
Please notice that the reliability of the results depends on the data
availability within each year. For details see \code{References}.

Arguments specifying \code{df} column names represent FLUXNET standard. To
process \code{REddyProc} outputs, timestamp must be corrected to represent
middle of averaging period and appropriate columns selected (see
\code{Examples}).
}
\section{References}{
  Griebel, A., Metzen, D., Pendall, E., Burba, G., &
  Metzger, S. (2020). Generating spatially robust carbon budgets from flux
  tower observations. Geophysical Research Letters, 47, e2019GL085942.
  https://doi.org/10.1029/2019GL085942
}

\examples{
\dontrun{
library(REddyProc)

# convert timestamp
DETha98 <- fConvertTimeToPosix(Example_DETha98, 'YDH', Year = 'Year',
Day = 'DoY', Hour = 'Hour')[-(2:4)]

# generate randomly wind directions for demonstration purposes (not included)
DETha98$WD <- sample(0:359, nrow(DETha98), replace = TRUE)

# if QcCol = NULL, all non-missing values of targetCol are used for budgeting
not_filled <- DETha98
not_filled$DateTime <- not_filled$DateTime - 900
spti_boot(not_filled, "DateTime", "WD", "LE", NULL, conv_fac = 1e-6)

# gap-filling is not needed but illustrates processing of FLUXNET data
# notice that ustar filtering of NEE should be applied before budgeting
DETha98 <- filterLongRuns(DETha98, "NEE")
EProc <- sEddyProc$new('DE-Tha', DETha98,
c('NEE', 'Rg', 'Tair', 'VPD', 'Ustar'))
EProc$sMDSGapFillAfterUstar('NEE', uStarTh = 0.3, FillAll = TRUE)
filled <- cbind(DETha98, EProc$sExportResults())

# correct timestamp to represent middle of averaging period
filled$DateTime <- filled$DateTime - 900
spti_boot(filled, "DateTime", "WD", "NEE", "NEE_uStar_fqc")
}
}
