% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_handling.R
\name{strip_suffix}
\alias{strip_suffix}
\title{Strip Positional Qualifier Suffix}
\usage{
strip_suffix(x, warn = FALSE)
}
\arguments{
\item{x}{A string vector.}

\item{warn}{A logical value}
}
\value{
A string vector with extracted variable codes.
}
\description{
Function removes from variable name the suffix with three indices
representing horizontal and vertical placement and number of replicates
(_H_V_R suffix used in tower network naming strategy).
}
\details{
If \code{warn = TRUE}, it is checked if multiple _H_V_R suffixes were
detected. This might be undesired based on the application.
}
\examples{
x <- c("TA_1_1_1", "TS_1_1_1", "VPD")
strip_suffix(x)
}
