% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_handling.R
\name{summary_QC}
\alias{summary_QC}
\title{Quality Control Summary}
\usage{
summary_QC(
  x,
  qc_names,
  cumul = FALSE,
  plot = FALSE,
  perc = TRUE,
  flux = NULL,
  na.as = NULL,
  na.as_0_pattern = "spikesLF$|fetch70$|man$",
  additive = NULL,
  additive_pattern = "interdep$|wresid$",
  no_messages = FALSE
)
}
\arguments{
\item{x}{A data frame with column names.}

\item{qc_names}{A vector of names of data frame \code{x} columns to combine.}

\item{cumul}{A logical value that determines if cumulative (\code{cumul =
TRUE}) or individual (\code{cumul = FALSE}) effects of quality control
flags should be shown.}

\item{plot}{A logical value. If \code{TRUE}, the results are represented as a
ggplot object. If \code{FALSE}, they are represented as a table.}

\item{perc}{A logical value. If \code{TRUE}, the results are reported in
percentages. If \code{FALSE}, counts are used instead.}

\item{flux}{A character string. Used only if \code{plot = TRUE}. Includes the
flux name in the plot title to emphasize the relevance of displayed quality
control filters.}

\item{na.as}{\code{NULL} or a vector of integer or \code{NA} values
determining interpretation of missing flags in each respective column of
\code{x} given by \code{qc_names}. If \code{NULL}, automated recognition is
used. If only one value is provided, all columns are treated the same way.}

\item{na.as_0_pattern}{A character string. A \code{\link[=regexp]{regular
expression}} \code{\link{grep}} \code{pattern} identifying \code{qc_names}
for which \code{NA} flags are interpreted as zeros.}

\item{additive}{\code{NULL} or a vector of logical values (\code{TRUE} or
\code{FALSE}) determining additivity of each respective column of \code{x}
given by \code{qc_names}. If \code{NULL}, automated recognition is used.
Otherwise, values determine if the flags should be treated as additive
(\code{additive = TRUE}) or with fixed effect (\code{additive = FALSE}). If
only one value is provided, all columns are considered to be of the same
type.}

\item{additive_pattern}{A character string. A \code{\link[=regexp]{regular
expression}} \code{\link{grep}} \code{pattern} identifying \code{qc_names}
of flags with additive effect.}

\item{no_messages}{A logical value.}
}
\value{
A table or a ggplot object depending on the \code{plot} argument
  value. If \code{length(qc_names) == 0}, \code{NULL} is returned instead.
}
\description{
\code{summary_QC} is a function that summarizes quality checking results in a
form of table or plot.
}
\details{
\code{summary_QC} loads a data frame \code{x}, extracts quality control (QC)
columns from it based on \code{qc_names} and creates a table (\code{plot =
FALSE}) or a plot (\code{plot = TRUE}) for these columns. Results are
displayed as percentages (\code{perc = TRUE}) or counts (\code{perc = FALSE})
for given flag and QC filter.

\code{cumul = TRUE} specifies that cumulative effect of gradually applied QC
filters on resulting flags is considered. Note that for \code{cumul = TRUE}
the results do depend on the order of qc_names. \code{additive} is considered
only if \code{cumul = TRUE}, otherwise skipped.

For a detailed description of automated recognition see
\code{\link{combn_QC}}.
}
\examples{
set.seed(5)
aa <- as.data.frame(replicate(
4, sample(c(0:2, NA), 20, replace = TRUE, prob = c(rep(0.3, 3), 0.1))))
names(aa) <- letters[1:4]

bb <- summary_QC(aa, letters[1:4])
is_add <- c(FALSE, TRUE, FALSE, TRUE)
summary_QC(aa, letters[1:4], na.as = c(NA, 0, NA, 2))
summary_QC(aa, letters[1:4], cumul = TRUE, additive = TRUE)
summary_QC(aa, letters[1:4], cumul = TRUE, additive = FALSE)
summary_QC(aa, letters[1:4], cumul = TRUE, additive = is_add)
cc <- summary_QC(aa, letters[1:4], cumul = FALSE, additive = is_add)
identical(bb, cc) # Argument additive is skipped when cumul = FALSE

library(ggplot2)
(xx <- summary_QC(aa, letters[1:4], cumul = TRUE, additive = TRUE,
plot = TRUE, flux = "CO2 flux"))
xx + ggplot2::theme(text = ggplot2::element_text(size = 20))

summary_QC(aa, rep(letters[1:4], 2),
cumul = TRUE, additive = TRUE, perc = FALSE)
summary_QC(aa, rep(letters[1:4], 2),
cumul = TRUE, additive = TRUE, plot = TRUE, perc = FALSE)
}
\seealso{
\code{\link{combn_QC}}, \code{\link{ggplot}}.
}
