% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quality_checking.R
\name{use_QC}
\alias{use_QC}
\title{Use Quality Control}
\usage{
use_QC(x, qc, qc_thr = 2, na.as = NA)
}
\arguments{
\item{x}{A numeric vector.}

\item{qc}{A numeric vector.}

\item{qc_thr}{A numeric value.}

\item{na.as}{A numeric value or \code{NA}.}
}
\value{
A numeric vector of length \code{x}.
}
\description{
Apply quality control information to the respective numeric vector.
}
\details{
Values of \code{x} are set to \code{NA} if \code{qc} value (flag) equals or
is above \code{qc_thr} or if \code{qc} value is missing.

\code{na.as = 0} can be used to preserve \code{x} values with respective
\code{NA} \code{qc} values.
}
\examples{
set.seed(215)
qc <- sample(0:2, 10, replace = TRUE)
use_QC(1:10, qc)

# set stricter qc_thr
use_QC(1:10, qc, qc_thr = 1)

# NAs in qc
is.na(qc) <- c(4, 6)
use_QC(1:10, qc)

# preserve x values for NA qc
use_QC(1:10, qc, na.as = 0)

}
