% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_handling.R
\name{write_eddy}
\alias{write_eddy}
\title{Eddy Covariance Data Output}
\usage{
write_eddy(x, file = "", append = FALSE, quote = TRUE, sep = ",",
  units_fill = "-", na = "-9999", row.names = FALSE, col.names = TRUE,
  qmethod = "double", ...)
}
\arguments{
\item{x}{The object that will be written. It is a data frame with optional
attributes \code{units} and \code{varnames} assigned to each column.
Otherwise it is converted by \code{\link{as.data.frame}}.}

\item{file}{Either a character string naming a file to write to or a
\code{\link{connection}} that is open for writing. \code{""} results in
writing to the console.}

\item{append}{A logical value. It is considered only if \code{file} is not a
\code{connection}. If \code{TRUE}, the output is written below the content
of the file. If \code{FALSE}, the content of the file is overwritten.}

\item{quote}{A logical value (\code{TRUE} or \code{FALSE}) or a numeric
vector. If \code{TRUE}, columns of class character or factor will be
surrounded by double quotes. If a numeric vector, its elements should mark
the indices of character or factor columns to quote. In both cases, row and
column names and units are quoted if present. If \code{FALSE}, no quoting
is performed.}

\item{sep}{A character used as the field separator of each row.}

\item{units_fill}{A character string that represents missing value of
\code{units} attribute in the output.}

\item{na}{A character string that represents missing data values in the
output.}

\item{row.names}{Either a logical value (\code{TRUE} or \code{FALSE}) that
determines if the row names of \code{x} should be included in the output,
or a character vector of row names that will be used instead.}

\item{col.names}{Either a logical value (\code{TRUE}, \code{FALSE} or
\code{NA}) or a character vector. If \code{TRUE}, column names of \code{x}
will be included in the output. If a character vector, its elements will be
used as column names. If \code{x} is supplied as vector, an attempt is made
to extract meaningful variable name from its attribute \code{varnames}. In
all cases, units extracted from \code{units} attribute of each column will
be written one line below column names with identical format. See the 'CSV
files' section in \code{\link{write.table}} for further explanation of
\code{col.names = NA}.}

\item{qmethod}{A character string. It determines the way how strings quoting
is performed in case of embedded double quote characters. The options are
either \code{"double"} (\code{write.csv} and \code{write.csv2} defaults),
that doubles the quote character, or \code{"escape"} (\code{write.table}
default), that escapes it in C style by a backslash.}

\item{...}{Further arguments to be passed to the internal
\code{\link{write.table}} function.}
}
\description{
Facilitates printing object \code{x} also with its units of measurement (or
space efficient metadata) to a file or \code{\link{connection}}.
}
\details{
\code{write_eddy} extends the possibilities of \code{write.table} so the
units of measurement can also be written. However, it uses default arguments
of \code{write.csv} to provide flexibility for the user and to accomodate the
function for the most common case. The character string \code{"-9999"} is
typically used to represent missing values in eddy covariance (\emph{eddy})
data.

Storing \code{varnames} and \code{units} attributes is practical mostly
within data frames and vectors. Attribute \code{varnames} extracted from each
data frame column represents names of respective variables and its main
purpose is to keep variable names of isolated vectors. Attribute \code{units}
extracted from each column represents units of measurement (or space
efficient metadata) of respective variables that are written one line above
data region. If the \code{varnames} or \code{units} attribute of given column
is \code{NULL}, of length not equal to 1, or contains missing value or empty
string, it is not considered meaningful. In that case the default column name
produced by \code{\link{as.data.frame}} is used instead (considered only if
\code{x} is supplied as vector) and unit of measurement is substituted with
\code{units_fill} string. \code{units_fill} can be an empty string.

Units of measurement are considered to be part of the output header and
therefore \code{col.names} and \code{quote} arguments have the effect on the
way they are written.
}
\examples{
xx <- read_eddy(text =
"timestamp,height,weight
\%d.\%m.\%Y,m,kg
24.1.2015,1.70,75
24.1.2016,1.72,78")
str(xx)
write_eddy(xx, "")

## NB: 'varnames' and 'units' attributes are dropped when you subset rows
## but unchanged if you subset columns:
write_eddy(xx[, 1], "") # 'varnames' attribute of the vector used as column name
write_eddy(head(xx), "") # dropped 'units' attribute

\dontrun{
# Example of using "col.names = NA"
zz <- file("ex.data", "w")  # open an output file connection
write_eddy(xx, zz, row.names = T, col.names = NA)
close(zz)
(ex_data <- read_eddy("ex.data", row.names = 1))
str(ex_data)
unlink("ex.data")}
}
\seealso{
\code{\link{write.table}} for information about full list of allowed
  arguments and their descriptions.

  \code{\link{read_eddy}} to read data frame with \code{varnames} and
  \code{units} attributes specified for each column.
}
