% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{choose_avail}
\alias{choose_avail}
\title{Choose Available Names from Set}
\usage{
choose_avail(names, all_names, show_ignored = FALSE, warn = TRUE)
}
\arguments{
\item{names}{A character vector with available names.}

\item{all_names}{A character vector with all expected variable names.}

\item{show_ignored}{A logical value. Should ignored names be shown?}

\item{warn}{A logical value. Should you be warned if \code{names} or
\code{all_names} contain duplicates?}
}
\value{
A character vector with subset of expected variable names.
}
\description{
Available variable names are checked against the full set of expected
variables and missing cases are reported.
}
\examples{
# all available names to choose from
all_names <- c("TA", "TS", "VPD", "LE", "H", "NEE")

# names to choose if present in all_names
# - "PM10" wanted but not available - thus reported as not available
# - c("TA", "TS", "VPD", "NEE") available but not wanted, thus ignored
names <- c("H", "LE", "PM10")
choose_avail(names, all_names)

}
