% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{coded_vars}
\alias{coded_vars}
\title{used within extract_QC() to link column names with SA and SAGA names}
\format{
An object of class \code{data.frame} with 8 rows and 4 columns.
}
\usage{
coded_vars
}
\description{
used within extract_QC() to link column names with SA and SAGA names
}
\keyword{internal}
