% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{combn_QC}
\alias{combn_QC}
\title{Combine Quality Checking Results}
\usage{
combn_QC(
  x,
  qc_names,
  name_out = "-",
  additive = NULL,
  na.as = NULL,
  additive_pattern = "interdep$|wresid$",
  na.as_0_pattern = "spikesLF$|fetch70$|man$",
  no_messages = FALSE
)
}
\arguments{
\item{x}{A data frame with column names.}

\item{qc_names}{A vector of names of data frame \code{x} columns to combine.}

\item{name_out}{A character string providing \code{varnames} value of the
output.}

\item{additive}{\code{NULL} or a vector of logical values (\code{TRUE} or
\code{FALSE}) determining additivity of each respective column of \code{x}
given by \code{qc_names}. If \code{NULL}, automated recognition is used.
Otherwise, values determine if the flags should be treated as additive
(\code{additive = TRUE}) or with fixed effect (\code{additive = FALSE}). If
only one value is provided, all columns are considered to be of the same
type.}

\item{na.as}{\code{NULL} or a vector of integer or \code{NA} values
determining interpretation of missing flags in each respective column of
\code{x} given by \code{qc_names}. If \code{NULL}, automated recognition is
used. If only one value is provided, all columns are treated the same way.}

\item{additive_pattern}{A character string. A \code{\link[=regexp]{regular
expression}} \code{\link{grep}} \code{pattern} identifying \code{qc_names}
of flags with additive effect.}

\item{na.as_0_pattern}{A character string. A \code{\link[=regexp]{regular
expression}} \code{\link{grep}} \code{pattern} identifying \code{qc_names}
for which \code{NA} flags are interpreted as zeros.}

\item{no_messages}{A logical value.}
}
\value{
An integer vector with attributes \code{varnames} and \code{units} is
  produced. \code{varnames} value is set by \code{name_out} argument. Default
  value of \code{varnames} and \code{units} is set to \code{"-"}.
}
\description{
Combine quality checking results depending on whether they have a fixed or
cumulative effect or any combination of these effects. It is also checked how
should \code{NA}s be interpreted.
}
\details{
The quality checking results to combine must be provided as columns of a data
frame \code{x}, optionally with any number of further columns that will be
ignored. Columns specified by \code{qc_names} will be further separated
according to their additivity. For flags with fixed effect (\code{additive =
FALSE}; the most typical type), maximum is taken over each row. For flags
with additive effect (\code{additive = TRUE}), sum is taken over each row. In
case both types of flags are present, results for both groups are summed
together.

The most typical value of argument \code{na.as} is \code{NA}. \code{NA} value
does not suggest any change in interpretation (value of variable
corresponding to this flag will be removed within quality checking scheme).
Exceptionally, value \code{0} can be used in case that the \code{NA} flag of
the quality checking test/filter is an expected result and means that the
half-hour was not checked by the given test/filter (e.g.
\code{\link{despikeLF}}).
}
\section{Automated recognition}{
 Default values for \code{additive} and
  \code{na.as} arguments are \code{FALSE} and \code{NA}, respectively. In
  case that \code{additive_pattern} is found within \code{qc_names} (i.e.
  \code{qc_names} ending with \code{"interdep"} or \code{"wresid"} pattern),
  respective values of \code{additive} are changed to \code{TRUE}. This is
  because \code{\link{interdep}} and wresid (see \code{\link{extract_QC}})
  quality control checks are defined as additive within the current quality
  control scheme. If \code{na.as_0_pattern} is detected within
  \code{qc_names} (i.e. \code{qc_names} ending with \code{"spikesLF"},
  \code{"fetch70"} or \code{"man"} pattern), respective values of
  \code{na.as} are changed to \code{0} (see \code{\link{despikeLF}}).
}

\examples{
set.seed(5)
aa <- data.frame(xx = sample(c(0:2, NA), 20, replace = TRUE))
aa$yy <- sample(c(0:2, NA), 20, replace = TRUE)
aa$add_F <- combn_QC(aa, qc_names = c("xx", "yy"), additive = FALSE,
name_out = "add_F")
aa$add_T <- combn_QC(aa, qc_names = c("xx", "yy"), additive = TRUE,
name_out = "add_T")
aa$add_F_na.as_0 <- combn_QC(aa, qc_names = c("xx", "yy"), additive = FALSE,
na.as = 0, name_out = "add_F_na.as_0")
aa$add_F_na.as_0part <- combn_QC(aa, qc_names = c("xx", "yy"),
additive = FALSE, na.as = c(0, NA), name_out = "add_F_na.as_0part")
aa$add_F_na.as_2 <- combn_QC(aa, qc_names = c("xx", "yy"), additive = FALSE,
na.as = 2, name_out = "add_F_na.as_2")
str(aa)
aa

}
\seealso{
\code{\link{summary_QC}}.
}
