% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_handling.R
\name{correct}
\alias{correct}
\title{Correct Character Vector}
\usage{
correct(x, attr = c("names", "units"), ...)
}
\arguments{
\item{x}{A character vector.}

\item{attr}{A character string identifying an attribute type a character
vector \code{x} to correct. Can be abbreviated.}

\item{...}{Further arguments to be passed to the internal
\code{\link{make.names}} function.}
}
\value{
A corrected character vector.
}
\description{
Substitute given characters or strings by their alternatives.
}
\details{
The function is intended to simplify the use of variable names and units
within the typical processing workflow employing
\code{\link[openeddy]{openeddy}}.

If \code{attr = "names"}, correction is meant to arrive to syntactically
valid names with a higher level of control. This assumes that the original
names were preserved during data loading (e.g. by using \code{check.names =
FALSE} in \code{\link{read_eddy}} or \code{\link{read.table}}). Specifically,
literal string \code{"(z-d)/L"} is renamed to \code{"zeta"} and specified
patterns or characters withing strings are substituted using regular
expression patterns: \itemize{\item \code{"co2_flux"} by \code{"NEE"} \item
\code{"*"} by \code{"star"} \item \code{"\%"} by \code{"perc"} \item
\code{"-"} and \code{"/"} by \code{"_"}.} After the substitutions
\code{make.names(names = x, ...)} is executed.

If \code{attr = "units"}, round and square brackets are substituted by an
empty string.
}
\examples{
correct(c("qc_co2_flux", "(z-d)/L", "x_70\%", "*[-(z-d)/L"))
correct(c("qc_co2_flux", "qc_NEE"), unique = TRUE)
correct(c("[m]", "(s)", "kg"), attr = "units")
}
\seealso{
\code{\link{make.names}}.
}
