% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quality_checking.R
\name{extract_QC}
\alias{extract_QC}
\title{Extract Quality Control Information}
\usage{
extract_QC(x, abslim = TRUE, spikesHF = TRUE, missfrac = TRUE,
  scf = TRUE, wresid = TRUE, rotation = c("double", "planar fit"),
  prefix = "[8]", split = "[/]", missfrac_thr = c(0.1, 0.1),
  scf_thr = c(2, 3), w_unrot_thr = c(0.35, 0.35), w_rot_thr = c(0.1,
  0.15))
}
\arguments{
\item{x}{A data frame with column names representing required variables.}

\item{abslim}{A logical value. Determines whether plausibility limits check
should be considered. If \code{abslim = TRUE}, column
\code{"absolute_limits_hf"} is required in \code{x}.}

\item{spikesHF}{A logical value. Determines whether check of spike percentage
in averaging period should be considered. If \code{spikesHF = TRUE}, column
\code{"spikes_hf"} is required in \code{x}.}

\item{missfrac}{A logical value. Determines whether check of missing data in
averaging period against thresholds should be done. If \code{missfrac =
TRUE}, columns \code{"file_records", "used_records", "w_spikes",
"u_spikes", "ts_spikes", "h2o_spikes" and "co2_spikes"} are required in
\code{x}.}

\item{scf}{A logical value. Determines whether check of spectral correction
factor against thresholds should be done. If \code{scf = TRUE}, columns
\code{"Tau_scf", "H_scf", "LE_scf" and "co2_scf"} are required in \code{x}.}

\item{wresid}{A logical value. Determines whether check of mean unrotated w
and w residual after planar fit against thresholds should be done. If
\code{wresid = TRUE}, columns \code{"w_unrot"} and \code{"w_rot"} are
required in \code{x}.}

\item{rotation}{A character string. Specifies the type of coordinate rotation
applied. Allowed values are "double" and "planar fit". Can be abbreviated.}

\item{prefix}{Character string containing a \code{\link{regular expression}}
identifying the prefix of coded values. See '\code{\link{extract_coded}}'.}

\item{split}{Character string containing a \code{\link{regular expression}}
identifying the separator of variable names in \code{units} attribute. See
'\code{\link{extract_coded}}'.}

\item{missfrac_thr}{A numeric vector with 2 non-missing values. Represents
thresholds (allowed fraction of missing high frequency data within
averaging period) used if \code{missfrac = TRUE}.}

\item{scf_thr}{A numeric vector with 2 non-missing values. Represents
thresholds used if \code{scf = TRUE}.}

\item{w_unrot_thr}{A numeric vector with 2 non-missing values. Represents
thresholds for unrotated w used if \code{wresid = TRUE}.}

\item{w_rot_thr}{A numeric vector with 2 non-missing values. Represents
thresholds for rotated w used if \code{wresid = TRUE}.}
}
\value{
A data frame. Each column has attributes \code{"varnames"} and
  \code{"units"} .
}
\description{
QC information stored in columns of data frame \code{x} is extracted and
interpreted in respect to instruments or individual fluxes.
}
\details{
The data frame \code{x} is expected to have certain properties. It is
required that it contains column names according to considered QC checks. See
'Arguments' above. Attribute \code{units} is required for columns
\code{"absolute_limits_hf" and "spikes_hf"} to extract the coded QC
information. See '\code{\link{extract_coded}}'.

Extracted QC information can be relevant to fluxes measured by given
instrument(s), specific flux or is applicable to all fluxes. See 'Naming
Strategy' below. Results are reported according to the QC scheme using QC
flag range 0 - 2. In cases when extracted variable is checked against
thresholds (missfrac, scf, wresid), \code{\link{apply_thr}} is used to assign
flag values. First value of \code{missfrac_thr}, \code{scf_thr},
\code{w_unrot_thr} or \code{w_rot_thr} sets threshold for flag 1, second
value sets threshold for flag 2. If both threshold values are the same, only
flag 0 and 2 will be resolved (hard flag).

Check of missing data in averaging period (missfrac) takes into account
number of valid records used for given averaging period. This number is
further reduced by the sum of count of high frequency data spikes out of
variables needed to compute covariance. Covariance pairs are w, u (Tau); w,
ts (H); w, h2o (LE) and w, co2 (NEE).
}
\section{Extracted QC Checks}{
 \itemize{ \item Check of plausibility limits
  (abslim). Test is not additive. \item Check of high frequency data spike
  percentage in averaging period against thresholds (spikesHF). Test is not
  additive. \item Check of missing data in averaging period against
  thresholds (missfrac). Test is not additive. \item Check of spectral
  correction factor against thresholds (scf). Test is not additive. \item
  Check of mean unrotated w (double rotation) or w residual (planar fit)
  against thresholds (wresid). Additive test.}
}

\section{Content and format of columns}{
 \itemize{ \item
  \code{"absolute_limits_hf"}: hard flags (passed or failed the test) for
  individual variables for absolute limits. Limits for each variable are set
  in the post-processing software which also reports the resulting flags in a
  coded format. See '\code{\link{extract_coded}}'. \item \code{"spikes_hf"}:
  hard flags (passed or failed the test) for individual variables for spike
  test. Threshold for maximum allowed percentage of spikes within averaging
  period for all variables is set in the post-processing software which also
  reports the resulting flags in a coded format. See
  '\code{\link{extract_coded}}'. \item \code{"file_records"}: number of valid
  records found in the raw file \item \code{"used_records"}: number of valid
  records used at given averaging period. This number can also contain high
  frequency spikes that should be excluded. \item \code{"u_spikes",
  "ts_spikes", "h2o_spikes" and "co2_spikes"}: number of high frequency
  spikes detected at given averaging period in respective variable. Values
  can be set to 0 if \code{"used_records"} already accounts for spikes. \item
  \code{"Tau_scf", "H_scf", "LE_scf" and "co2_scf"}: spectral correction
  factor for given flux. Values are above 1. \item \code{"w_unrot", "w_rot"}:
  unrotated and rotated w wind component, respectively (should be close to
  0).}
}

\section{Naming Strategy}{
 \strong{QC prefixes} (specifies which flux is
  affected by that QC output): \itemize{ \item qc_SA: applicable to fluxes
  relying only on SA (Tau, H) \item qc_SA_IRGA: applicable to fluxes relying
  both on SA and IRGA (LE, NEE) \item qc_Tau, qc_H, qc_LE, qc_NEE: only
  applicable for the respective flux \item qc_ALL: applicable to all fluxes}

  \strong{QC suffixes} (specifies which QC check was applied to get this QC
  output): \itemize{ \item abslim, spikesHF, missfrac, wresid. See 'Included
  QC Checks' above.}
}

\section{Abbreviations}{
 \itemize{ \item QC: Quality Control \item SA: Sonic
  Anemometer \item IRGA: InfraRed Gas Analyzer \item Tau: Momentum flux [kg
  m-1 s-2] \item H: Sensible heat flux [W m-2] \item LE: Latent heat flux [W
  m-2] \item NEE: Net ecosystem exchange [umol m-2 s-1] \item u: Longitudinal
  wind speed component [m s-1] \item w: Vertical wind speed component [m s-1]
  \item ts: Sonic temperature [degC] \item h2o: H2O concentration [mmol
  mol-1] \item co2: CO2 concentration [umol mol-1]}.
}

\section{References}{
 Foken, T., Wichura, B., 1996. Tools for quality
  assessment of surface-based flux measurements. Agric. For. Meteorol. 78,
  83–105. doi:10.1016/0168-1923(95)02248-1

  Mauder, M., Cuntz, M., Drue, C., Graf, A., Rebmann, C., Schmid, H.P.,
  Schmidt, M., Steinbrecher, R., 2013. A strategy for quality and uncertainty
  assessment of long-term eddy-covariance measurements. Agric. For. Meteorol.
  169, 122-135. doi:10.1016/j.agrformet.2012.09.006
}

\seealso{
\code{\link{extract_coded}} and \code{\link{apply_thr}}.
}
