% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_checking.R
\name{interdep}
\alias{interdep}
\title{Flux Interdependency}
\usage{
interdep(qc_LE, qc_H = NULL, IRGA = c("en_closed", "open"))
}
\arguments{
\item{qc_LE}{An atomic type containing numeric or NA values. Combination of
all available QC flags for LE.}

\item{qc_H}{An atomic type containing numeric or NA values. Combination of
all available QC flags for H. Used only if \code{IRGA = "open"}.}

\item{IRGA}{A character string. Specifies the type of IRGA. Allowed values
are \code{"en_closed"} both for closed and enclosed path systems and
\code{"open"} for open path systems. Can be abbreviated.}
}
\value{
A data frame. Each column has attributes \code{"varnames"} and
\code{"units"}.
}
\description{
Interdependency of H, LE and NEE QC flags due to corrections/conversions.
}
\details{
Flux interdependency is an additive QC flag correction. Results follow the QC
scheme using QC flag range 0 - 2. Returned data frame follows the 'Naming
Strategy' described in \code{\link[=extract_QC]{extract_QC()}}. Returned QC flags have QC
suffix interdep.

To convert buoyancy flux to sensible heat flux (SND or Schotanus correction),
reliable measurements of LE must be available. To correct LE and NEE
estimated by open-path IRGA (\code{IRGA = "open"}) for the effects of density
fluctuations due to temperature and humidity fluctuations (WPL or Webb
correction), reliable measurements of H must be available. To perform WPL
correction of NEE estimated with any kind of IRGA, reliable measurements of
LE must be available. Thus following set of rules apply:

If \code{IRGA = "en_closed"}
\itemize{
\item If \code{qc_LE == 2 | is.na(qc_LE) == TRUE}: qc_H and qc_NEE flags are increased
by 1.
}

If \code{IRGA = "open"}
\itemize{
\item if \code{qc_LE == 2 | is.na(qc_LE) == TRUE}: qc_H flags are increased by 1.
\item If \code{qc_H == 2 | is.na(qc_H) == TRUE}: qc_LE flags are increased by 1.
\item If \code{qc_H == 2 | is.na(qc_H) == TRUE | qc_LE == 2 | is.na(qc_LE) == TRUE}:
qc_NEE flags are increased by 1.
}
}
\section{Abbreviations}{

\itemize{
\item QC: Quality Control
\item H: Sensible heat flux [W m-2]
\item LE: Latent heat flux [W m-2]
\item NEE: Net ecosystem exchange [umol m-2 s-1]
\item IRGA: Infrared Gas Analyzer.
}
}

\section{References}{
 Mauder, M., Cuntz, M., Drue, C., Graf, A., Rebmann, C.,
Schmid, H.P., Schmidt, M., Steinbrecher, R., 2013. A strategy for quality
and uncertainty assessment of long-term eddy-covariance measurements.
Agric. For. Meteorol. 169, 122-135.
\url{https://doi.org/10.1016/j.agrformet.2012.09.006}
}

\seealso{
\code{\link[=combn_QC]{combn_QC()}} and \code{\link[=extract_QC]{extract_QC()}}.
}
