% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_handling.R
\name{merge_eddy}
\alias{merge_eddy}
\title{Merge Data Frames on Regular Date-Time Sequence}
\usage{
merge_eddy(
  x,
  start = NULL,
  end = NULL,
  check_dupl = TRUE,
  freq = NULL,
  format = "\%Y-\%m-\%d \%H:\%M",
  tz = "GMT"
)
}
\arguments{
\item{x}{List of data frames, each with \code{"timestamp"} column of class
\code{"POSIXt"}. Optionally with attributes \code{varnames} and
\code{units} for each column.}

\item{start, end}{A \code{"POSIXt"} value or character string containing
date-time information in given \code{format} specifying first (last) value
of the generated timestamp.}

\item{check_dupl}{A logical value specifying whether rows with duplicated
date-time values checked across \code{x} elements should be excluded before
merging.}

\item{freq}{A numeric value specifying the frequency (in seconds) of the
generated timestamp.}

\item{format}{A character string. The default \code{format} is
\code{"\%Y-\%m-\%d \%H:\%M"}.}

\item{tz}{A time zone (see \code{\link{time zones}}) specification to be used
for the conversion.}
}
\value{
A data frame with attributes \code{varnames} and \code{units} for
  each column, containing date-time information in column \code{"timestamp"}.
}
\description{
Single or multiple data frames are merged either on existing or generated
date-time sequence.
}
\details{
The primary purpose of \code{merge_eddy} is to combine chunks of data
vertically along their column \code{"timestamp"} with date-time information.
This \code{"timestamp"} is expected to be regular with given
\code{freq}uency. Resulting data frame contains added rows with expected
date-time values that were missing in \code{"timestamp"} column, followed by
\code{NA}s. In case that \code{check_dupl = TRUE} and \code{"timestamp"}
values across \code{x} elements overlap, detected duplicated rows are removed
(the order in which duplicates are evaluated depends on the order of \code{x}
elements). A special case when \code{x} has only one element allows to fill
missing date-time values in \code{"timestamp"} column of given data frame.


The list of data frames, each with column \code{"timestamp"}, is sequentially
\code{\link{merge}}d using \code{\link{Reduce}}. A \emph{(full) outer join},
i.e. \code{merge(..., all = TRUE)}, is performed to keep all columns of
\code{x} elements. The order of \code{x} elements can affect the result.
Duplicated column names within \code{x} elements are corrected using
\code{\link{make.unique}}. The merged data frame is then merged on the
validated \code{"timestamp"} column that can be either automatically
extracted from \code{x} or manually specified.

For horizontal merging (adding columns instead of rows) \code{check_dupl =
FALSE} must be set but simple \code{\link{merge}} could be preferred.
Combination of vertical and horizontal merging should be avoided as it
depends on the order of \code{x} elements and can lead to row duplication.
Instead, data chunks from different data sources should be first separately
vertically merged and then merged horizontally in a following step.
}
\seealso{
\code{\link{merge}}, \code{\link{Reduce}}, \code{\link{strptime}},
  \code{\link{time zones}}, \code{\link{make.unique}}
}
