% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openeddy-package.R
\docType{package}
\name{openeddy-package}
\alias{openeddy}
\alias{openeddy-package}
\title{openeddy: Post-process eddy covariance data with ease}
\description{
The openeddy package is a software infrastructure for eddy covariance data post-processing. It provides tools for handling data with attached metadata, setting up quality control scheme and summarizing and plotting the results. It is aligned with REddyProc package that provides methods for uStar-filtering, gap-filling, and flux-partitioning. Thus the combined use of openeddy and REddyProc allows to run the whole eddy covariance post-processing chain. Learn more at \url{https://github.com/lsigut/EC_workflow}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lsigut/openeddy}
  \item Report bugs at \url{https://github.com/lsigut/openeddy/issues}
}

}
\author{
\strong{Maintainer}: Ladislav Sigut \email{sigut.l@czechglobe.cz}

Other contributors:
\itemize{
  \item Matthias Mauder \email{matthias.mauder@kit.edu} [contributor]
  \item Milan Fischer \email{fischer.milan@gmail.com} [contributor]
  \item Thomas Wutzler \email{twutz@bgc-jena.mpg.de} [contributor]
  \item Pavel Sedlak \email{sedlak@ufa.cas.cz} [contributor]
  \item Lenka Mzourkova Macalkova \email{mzourkova.l@gmail.com} [contributor]
  \item Radek Czerny \email{czerny.r@czechglobe.cz} [contributor]
  \item Ryan Patrick McGloin \email{mcgloin.r@czechglobe.cz} [contributor]
  \item Jiri Dusek \email{dusek.j@czechglobe.cz} [contributor]
  \item Natalia Kowalska \email{kowalska.n@czechglobe.cz} [contributor]
  \item Georg Jocher \email{jocher.g@czechglobe.cz} [contributor]
  \item Lenka Foltynova \email{foltynova.l@czechglobe.cz} [contributor]
  \item Marian Pavelka \email{pavelka.m@czechglobe.cz} [contributor]
  \item Gerardo Fratini \email{Gerardo.Fratini@licor.com} [contributor]
  \item Global Change Research Institute CAS [copyright holder]
  \item University of Ostrava [copyright holder]
}

}
\keyword{internal}
