% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openeddy.R
\docType{package}
\name{openeddy}
\alias{openeddy}
\alias{openeddy-package}
\title{The R Package for Low Frequency Eddy Covariance Data Processing}
\description{
This package provides utilities for eddy covariance data handling, quality
checking (similar to Mauder et al., 2013), processing, summarising and
plotting. It aims to standardise the automated quality checking and make data
processing reproducible. In combination with \code{REddyProc} package it
allows to process eddy covariance data with single processing chain. An
extended example describing the intended eddy covariance data processing
workflow is available at: \url{https://github.com/lsigut/EC_workflow}.
}
\section{References}{
 Mauder, M., Cuntz, M., Drue, C., Graf, A., Rebmann, C.,
  Schmid, H.P., Schmidt, M., Steinbrecher, R., 2013. A strategy for quality
  and uncertainty assessment of long-term eddy-covariance measurements.
  Agric. For. Meteorol. 169, 122-135. doi:10.1016/j.agrformet.2012.09.006
}

