% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_eddy.R
\name{plot_eddy}
\alias{plot_eddy}
\title{Plot Time Series of Eddy Covariance Data}
\usage{
plot_eddy(x, flux, qc_flag = "none", test = "none", flux_gf = "none",
  NEE_sep = FALSE, skip = c("none", "monthly", "weekly"),
  panel_top = c("T_light", "VPD_Rn", "H_err_var"),
  panel_bottom = c("VPD_Rn", "T_light", "H_err_var"), light = c("PAR",
  "GR"), GPP_check = TRUE, document = TRUE)
}
\arguments{
\item{x}{A data frame with column names representing required variables. See
'Details' below.}

\item{flux}{A character string. Specifies the column name in \code{x} with
flux values.}

\item{qc_flag}{A character string. Specifies the column name in \code{x} with
flux related quality control flag used for visualisation of data quality
and setting of y-axis range. If "none" is provided, all data will be used.
See 'Quality Control'.}

\item{test}{A character string. Specifies the column name in \code{x} with
quality control flag for visualisation of its effect on the data. If "none"
is provided, no visualisation will be done. See 'Quality Control'.}

\item{flux_gf}{A character string. Specifies the column name in \code{x} with
gap-filled flux values.}

\item{NEE_sep}{A logical value. Determines whether NEE separation should be
visualized. If \code{TRUE}, columns \code{"Reco"} and \code{"GPP"} are
expected in \code{x}.}

\item{skip}{A character string. Determines whether plotting should be done in
monthly (\code{skip = "weekly"}), weekly intervals (\code{skip =
"monthly"}) or both (\code{skip = "none"}).}

\item{panel_top}{A character string. Selects one of the available modules for
plotting additional variables. This module is displayed above the panel
with fluxes in weekly plots. Can be abbreviated. See 'Modules'.}

\item{panel_bottom}{A character string. Selects one of the available modules
for plotting additional variables. This module is displayed below the panel
with fluxes in weekly plots. Can be abbreviated. See 'Modules'.}

\item{light}{A character string. Required only for the \code{"T_light"}
module. Selects preferred variable for incoming light intensity.
\code{"PAR"} or \code{"GR"} is allowed. Can be abbreviated.}

\item{GPP_check}{A logical value. If \code{TRUE}, column \code{"GPP"} is
checked to be of the same sign convention as \code{"NEE"} column and
corrected in the case it is not. Required only if \code{NEE_sep = TRUE}.}

\item{document}{A logical value. If \code{TRUE}, values of \code{qc_flag} and
\code{test} arguments are documented in both monthly and weekly plots.}
}
\description{
Visualize flux measurements together with micrometeorological or other
related variables in monthly and weekly intervals. Missing values, scales of
axes and plotting regions are treated in an automated way.
}
\details{
The data frame \code{x} is expected to have certain properties. It is
required that it contains column named \code{"timestamp"} of class
\code{"POSIXt"} with regular sequence of date-time values, typically with
(half-)hourly frequency. Any missing values in \code{"timestamp"} are not
allowed. Thus, if no records exist for given date-time value, it still has to
be included. It also has to contain required (depends on the argument values
and applied modules) column names. If required auxiliary variable is not
available (e.g. \code{"P"}, \code{"Tsoil"}), it still has to be included in
\code{x} as a named column with \code{NA} values. The \code{x} column defined
by argument \code{flux} is the only one that has to contain at least one
non-missing value.

If \code{skip = "weekly"}, minimum requirements for \code{x} column names are
\code{"timestamp"} and \code{flux}. If \code{skip = "none"} or
\code{"monthly"}, \code{"P"} and respective names specified in 'Modules' (see
below) are also required.

Variable names for plot labels are extracted from required column names of
\code{x}. Units are extracted from \code{x} if they are present as
\code{units} attributes of required columns. If missing, \code{"-"} string is
used instead.

Plotting is separated into two stages. Firstly, \code{flux} time-series data
are drawn in monthly intervals. Monthly plotting regions consist of four
figures on top of each other representing separately four consecutive months.
Secondly, if \code{skip = "none"} or \code{"monthly"}, \code{flux}
time-series data are drawn together with auxiliary data in weekly intervals.
Weekly plotting regions are described in 'Modules' section (see below).
}
\section{Modules}{
 Applies only if \code{skip = "none"} or \code{"monthly"}.
  Plotting in weekly intervals is simplified by using predefined modules.
  Their main purpose is to achieve well-arranged display of auxiliary
  variables. Weekly plotting regions consist of two figures representing
  separately two consecutive weeks. Each figure contains three panels on top
  of each other. The middle panel always contains the values from \code{flux}
  and \code{"P"} columns of \code{x}. Variables used in the upper and lower
  panel can be changed by \code{panel_top} and \code{panel_bottom}. These
  arguments specify the respective modules that will be loaded (can be the
  same) and thus also a certain set of required column names of \code{x}
  (variables).

  Available modules are: \itemize{ \item T_light: requires \code{"Tair"},
  \code{"Tsoil"} and selected \code{light} columns. \item VPD_Rn: requires
  \code{"VPD"} and \code{"Rn"} columns. \item H_err_var: requires
  \code{"rand_err_H"} and \code{"ts_var"} columns.}
}

\section{Quality Control}{
 \code{qc_flag} and \code{test} relate to QC flags
  available for the specified \code{flux}. Only QC scheme using QC flag range
  0 - 2 is supported.

  Flags specified by \code{qc_flag} separate corresponding flux values to two
  groups. \emph{Used data} for flags 0 and 1 (black points) and
  \emph{Excluded data} for flags 2 and \code{NA} (grey points). The range of
  y-axis limit in the figures with \code{flux} values is based on \emph{Used
  data} only. If \code{qc_flag = "none"}, all data are \emph{Used data}. The
  time interval used for setting the range is one month both for monthly
  (respective month) and weekly (month centered on the respective week)
  plots.

  In order to emphasize \code{flux} values with lower quality, \code{test}
  can be specified. Values with QC flag = 1 have green center of the point.
  Values with QC flag = 2 or \code{NA} have red center of the point.

  NB: \code{flux} data with \code{NA} values are always \emph{Excluded data}
  and cannot be emphasized (\code{NA} values are not drawn).
}

\section{Gap-filling and NEE separation}{
 Gap-filled flux values can be
  displayed using \code{flux_gf} as a line overlaying the \code{flux} values.
  If \code{NEE_sep = TRUE}, columns \code{"Reco"} and \code{"GPP"} are
  expected in \code{x}. \code{GPP_check} is evaluated only if \code{NEE_sep =
  TRUE}. If \code{GPP_check = TRUE}, mean GPP is checked to be negative so it
  compares well with NEE and corrected accordingly if needed. Thus negative
  values denote carbon sink. Such test can fail for measurements above
  surfaces without vegetation or ecosystems with insignificant sink. In that
  case set \code{GPP_check = FALSE} and check and correct GPP sign convention
  manually.
}

\section{Abbreviations}{
 \itemize{ \item H: Sensible heat flux [W m-2] \item
  NEE: Net Ecosystem Exchange [umol m-2 s-1] \item GPP: Gross Primary
  Production [umol m-2 s-1] \item Reco: Ecosystem Respiration [umol m-2 s-1]
  \item QC: Quality Control \item P: Precipitation [mm] \item PAR:
  Photosynthetic Active Radiation [umol m-2 s-1] \item GR: Global Radiation
  [W m-2] \item T: Temperature [degC] \item Tair: Air Temperature [degC]
  \item Tsoil: Soil Temperature [degC] \item VPD: Vapor Pressure Deficit
  [hPa] \item Rn: Net Radiation [W m-2] \item rand_err_H: random error of H
  [W m-2]; in plots abbreviated as H_re \item ts_var: sonic temperature
  variance [K2]}
}

\seealso{
\code{\link{read_eddy}} and \code{\link{strptime_eddy}}.
}
