% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_handling.R
\name{remap_vars}
\alias{remap_vars}
\title{Remap Variables}
\usage{
remap_vars(x, new, source, regexp = FALSE, qc = NULL, na.rm = TRUE)
}
\arguments{
\item{x}{data frame}

\item{new}{A character vector of new column names for remapping.}

\item{source}{A vector of \code{x} column names matching \code{new} to remap.
If \code{regexp = TRUE}, character vector containing \code{\link{regular
expression}}s.}

\item{regexp}{A logical value. If \code{FALSE} (the default), \code{source}
will be interpreted literary. If \code{TRUE}, \code{source} elements will
be used as \code{\link{grep}} \code{pattern}s.}

\item{qc}{A character string. A \code{\link{regular expression}}
\code{\link{grep}} \code{pattern} identifying \code{x} column names that
carry quality control information for respective \code{source}.}

\item{na.rm}{A logical value indicating whether \code{NA} values should be
stripped before the computation proceeds. \code{na.rm} is used only if
\code{regexp = TRUE} and multiple columns identified by \code{source} are
combined by averaging.}
}
\value{
A data frame with attributes varnames and units assigned to each
  respective column.
}
\description{
Extracts and renames specified columns of a data frame, computes mean in case
of \code{\link{regular expression}} pattern matching multiple column names or
initializes one if missing.
}
\details{
New data frame is created based on \code{x} and specified \code{source}.
Original \code{x} names are changed according to respective \code{new}
elements and kept as \code{varnames} attributes for tracebility. Accordingly,
if \code{qc} is specified, quality control columns are marked by \code{"qc_"}
prefix.

If \code{regexp = FALSE} (the default), strictly one variable (column) will
be remapped to new name. The \code{source} elements must exactly match
\code{x} names, otherwise expected column is initialized with \code{NA}s. If
\code{qc} is specified, strictly one respective quality control column will
be renamed or skipped if not present.

If \code{regexp = TRUE}, multiple columns can match the \code{source} element
\code{\link{regular expression}} pattern. In that case \code{\link{rowMeans}}
are produced and names of averaged columns kept as \code{varnames} attributes
for tracebility. Similarly, also quality control flags are averaged over
available columns if \code{qc} is specified. Note that variable names need to
have unique patterns in order to achieve expected results. E.g. precipitation
abbreviated as P will have overlap with PAR; instead, sumP can be used.

\code{varnames} attribute is expected. If not automatically assigned to
\code{x} through \code{\link{read_eddy}} when read from a file, they should
be assigned before remapping to keep documentation (especially if multiple
columns are combined to a single one).
}
\seealso{
\code{\link{varnames}}.
}
