% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_checking.R
\name{separate}
\alias{separate}
\title{Separate columns}
\usage{
separate(x, ...)
}
\description{
Helper function utilized in \code{\link{extract_coded}} for coded variables
that controls column splitting.
}
\keyword{internal}
