% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_handling.R
\name{set_OT_input}
\alias{set_OT_input}
\title{Create Input for REddyProc/Online Tool}
\usage{
set_OT_input(x, names_in, names_out = c("qcNEE", "NEE", "qcLE", "LE", "qcH",
  "H", "Rg", "Tair", "Tsoil", "rH", "VPD", "qcTau", "Ustar"), qcTau = TRUE,
  qcH = TRUE, qcLE = TRUE, qcNEE = TRUE, check_time = TRUE,
  check_VPD = TRUE)
}
\arguments{
\item{x}{A data frame with column names and \code{"timestamp"} column in
POSIXt format.}

\item{names_in}{Column names (variables) present in \code{x} that will be
used as input.}

\item{names_out}{Column names required by the tools for respective
\code{names_in}.}

\item{qcTau}{A logical value indicating whether quality control of momentum
flux should be considered.}

\item{qcH}{A logical value indicating whether quality control of sensible
heat flux should be considered.}

\item{qcLE}{A logical value indicating whether quality control of latent heat
flux should be considered.}

\item{qcNEE}{A logical value indicating whether quality control of net
ecosystem exchange should be considered.}

\item{check_time}{A logical value indicating whether timestamp should be
checked.}

\item{check_VPD}{A logical value indicating whether range of vapor pressure
deficit values should be checked.}
}
\description{
Creates input for gap-filling and flux partitioning tools implemented either
offline in R (REddyProc) or accessible online
(\href{http://www.bgc-jena.mpg.de/~MDIwork/eddyproc/upload.php}{Online Tool})
from the data frame \code{x}. Columns of data frame \code{x} ideally have
assigned attributes \code{varnames} and \code{units}.
}
\details{
The typical variables (column names; i.e. \code{names_out}) required by the
tools are quality control of net ecosystem exchange (\code{"qcNEE"}), net
ecosystem exchange (\code{"NEE"}), quality control of latent heat
(\code{"qcLE"}), latent heat (\code{"LE"}), quality control of sensible heat
(\code{"qcH"}), sensible heat (\code{"H"}), global radiation (\code{"Rg"}),
air temperature (\code{"Tair"}), soil temperature (\code{"Tsoil"}), relative
humidity (\code{"rH"}), quality control of momentum flux (\code{"qcTau"}) and
friction velocity (\code{"Ustar"}). Check the gap-filling tool documentation
for more details.

Arguments \code{qcTau}, \code{qcH}, \code{qcLE} and \code{qcNEE} determine
whether the quality control will be applied to the respective fluxes. In case
of \code{qcTau}, quality control is applied to friction velocity (Ustar). If
\code{TRUE}, values of fluxes or friction velocity are set to \code{NA} when
respective quality control flag is \code{NA} or higher than \code{1}. In case
of \code{qcNEE}, NEE is also set to \code{NA} if respective values of Ustar
are \code{NA} (after application \code{qcTau} argument). This conservative
approach will assure that NEE values that cannot be compared against friction
velocity threshold (Ustar filtering) will be excluded.

\code{check_time} checks the timestamp minutes that should be in format
specifying the end of measurement interval, i.e. [0, 30] instead of [15, 45].
Otherwise it produces warning. This check is designed only for data reported
in half-hourly intervals.

\code{check_VPD} checks that the range of vapor pressure deficit (VPD) values
complies with assumption VPD <= 100. This is to check that VPD units are hPa,
not Pa.
}
\seealso{
\code{\link{read_eddy}} and \code{\link{write_eddy}}.
}
