% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_handling.R
\name{strptime_eddy}
\alias{strptime_eddy}
\title{Conversion of Regular Date-time Sequence from Character}
\usage{
strptime_eddy(
  x,
  format = "\%Y-\%m-\%d \%H:\%M",
  freq = 1800,
  shift.by = NULL,
  tz = "GMT",
  ...
)
}
\arguments{
\item{x}{A character vector containing date-time information to be converted
to class \code{"POSIXct"}.}

\item{format}{A character string. The default \code{format} is
\code{"\%Y-\%m-\%d \%H:\%M"}}

\item{freq}{A numeric value specifying the frequency (in seconds) of the
input date-time vector.}

\item{shift.by}{A numeric value specifying the time shift (in seconds) to be
applied to the date-time information.}

\item{tz}{A time zone (see \code{\link{time zones}}) specification to be used
for the conversion.}

\item{...}{Further arguments to be passed from or to other methods.}
}
\description{
Converts character vector to class \code{"POSIXct"} using
\code{\link{strptime}} and validates the result. The input has to represent a
regular date-time sequence with given frequency. Additional attributes
\code{varnames} and \code{units} are assigned to returned vector with fixed
strings \code{"timestamp"} and \code{"-"}, respectively.
}
\details{
Eddy covariance related measurements are usually stored with a timestamp
representing the end of the averaging period (typicaly 1800 s) in standard
time. This can however cause difficulties during data aggregation or
plotting. Therefore it is recommended to shift the date-time information
using \code{shift.by} to represent the center of averaging period prior to
any computations. It is also recommended to change the date-time information
to its original state before saving to a file (see Examples section).

Any unsuccessful attempt to convert date-time information is considered to be
unexpected behaviour and returns an error message instead of \code{NA} value.
In case that multiple formats are present in the timestamp, it has to be
corrected prior using \code{strptime_eddy}. It is expected that time series
are continuous although no valid measurements are available for given time
interval. Therefore \code{freq} value is checked against the lagged
differences (\code{\link{diff}}) applied to the converted date-time vector
and returns an error message if mismatch is found.
}
\examples{
xx <- c("01.01.2014  00:30:00", "01.01.2014  01:00:00",
"01.01.2014  01:30:00", "01.01.2014  02:00:00")
varnames(xx) <- "timestamp"
units(xx) <- "-"
str(xx)
(yy <- strptime_eddy(xx, "\%d.\%m.\%Y \%H:\%M", shift.by = -900))
## Convert to original format
format(yy + 900, format = "\%d.\%m.\%Y \%H:\%M", tz = "GMT")
attributes(yy)
\dontrun{
zz <- xx[-3]
## This is not a regular date-time sequence
strptime_eddy(zz, "\%d.\%m.\%Y \%H:\%M")
## freq argument provided incorrectly
strptime_eddy(xx, "\%d.\%m.\%Y \%H:\%M", freq = 3600)}
}
\seealso{
\code{\link{strptime}} provides the details about conversions
  between date-time character representation and \code{"POSIXct"} or
  \code{"POSIXlt"} classes. It also includes information about \code{format}
  \emph{conversion specification}.

  \code{\link{DateTimeClasses}} further inform about the date-time classes.

  See \code{\link{locales}} to query or set a locale.
}
