% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{structure_eddy}
\alias{structure_eddy}
\title{Folder Structure Setup}
\usage{
structure_eddy(root = ".", create_dirs = FALSE, fsep = .Platform$file.sep, ...)
}
\arguments{
\item{root}{A character string defining the root of created folder structure.}

\item{create_dirs}{A logical value. Indicates whether directories should be
created.}

\item{fsep}{A character. The path separator to use (assumed to be ASCII).}

\item{...}{Further arguments to be passed to \code{dir.create} function.}
}
\value{
A named list with paths to folder structure directories.
Corresponding directories are created as a function side effect if
\code{create_dirs = TRUE}.
}
\description{
Folder structure recommended for eddy covariance data processing and
archivation.
}
\details{
The purpose is to standardize the locations for data and metadata used in the
proposed workflow (\url{https://github.com/lsigut/EC_workflow}) and separate
them into levels corresponding to the processing stage. The folder structure
is not required to successfully apply the workflow but simplifies its use.

Data processing stages \itemize{
\item level_1: EddyPro full output files, meteo data and their merged product
as an input for quality checking.
\item level_2: quality checking results and documentation, and files used as
inputs for gap-filling.
\item level_3: gap-filling output and its documentation, summary of the
computed fluxes and meteorological data including their aggregation and
plotting.}
}
\examples{
xx <- structure_eddy()
xx
xx$Input_for_GF

}
\seealso{
\code{\link{file.path}}
}
