% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_handling.R
\name{structure_eddy}
\alias{structure_eddy}
\title{Folder Structure Setup}
\usage{
structure_eddy(root = ".", create_dirs = FALSE, fsep = .Platform$file.sep, ...)
}
\arguments{
\item{root}{A character string defining the root of created folder structure.}

\item{create_dirs}{A logical value. Indicates whether directories should be
created.}

\item{fsep}{A character. The path separator to use (assumed to be ASCII).}

\item{...}{Further arguments to be passed to \code{dir.create} function.}
}
\value{
A named list with paths to folder structure directories.
  Corresponding directories are created as a function side effect if
  \code{create_dirs = TRUE}.
}
\description{
Folder structure recommended for eddy covariance data processing and
archivation.
}
\details{
The purpose is to standardize the locations for metadata and post-processing
inputs required to run the proposed workflow
(\url{https://github.com/lsigut/EC_workflow}) as well as to store data and
metadata in levels corresponding to processing stage. The folder structure is
not required to successfully apply the workflow but simplifies its use.

Data processing stages \itemize{\item Level 0: Raw files with measured high
frequency eddy covariance data and relevant metadata or instrument setup
files. \item Level 1: Processing software setup and output files and a
logbook. \item Level 2: Quality checking results and documentation,
definition of ecosystem boundary, storage flux processing and files used as
inputs for Level 3 data. \item Level 3: gap-filling output and its
documentation, summary of the computed fluxes and meteorological data
including their aggregation.}
}
