% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_handling.R
\name{summary_QC}
\alias{summary_QC}
\title{Quality Control Summary}
\usage{
summary_QC(
  x,
  qc_names,
  na.as = NA,
  cumul = FALSE,
  additive = FALSE,
  plot = FALSE,
  perc = TRUE,
  flux = NULL
)
}
\arguments{
\item{x}{A data frame with column names.}

\item{qc_names}{A vector of names of data frame columns to combine.}

\item{na.as}{A vector of numeric or \code{NA} values for the \code{qc_names}
subset of \code{x} that determines how should be the missing flags
interpreted. If only one value is provided, all columns are treated the
same way.}

\item{cumul}{A logical value that determines if cumulative (\code{cumul =
TRUE}) or individual (\code{cumul = FALSE}) effects of quality control
flags should be shown.}

\item{additive}{A vector of logical values (\code{TRUE} or \code{FALSE}) for
the \code{qc_names} subset of \code{x} that determines if the flags should
be treated as additive (\code{additive = TRUE}) or with fixed effect
(\code{additive = FALSE}). If only one value is provided, all columns are
considered to be of the same type.}

\item{plot}{A logical value. If \code{TRUE}, the results are plotted in the
active graphical device. If \code{FALSE}, they are represented as a table.}

\item{perc}{A logical value. If \code{TRUE}, the results are reported in
percentages. If \code{FALSE}, counts are used instead.}

\item{flux}{A character string. Used only if \code{plot = TRUE}. Includes the
flux name in the plot title to emphasise the relevance of displayed test
types.}
}
\value{
A table or a plot depending on the \code{plot} argument value.
}
\description{
\code{summary_QC} is a function that summarizes quality checking results in a
form of table or plot.
}
\details{
\code{summary_QC} loads a data frame \code{x}, extracts quality control
columns from it based on \code{qc_names} and creates a table (\code{plot =
FALSE}) or a plot (\code{plot = TRUE}) for these columns. Results are
displayed in percentage (\code{perc = TRUE}) or counts (\code{perc = FALSE})
of given flag per dataset.

\code{cumul = TRUE} specifies that cumulative effect of flags is considered
for each halfhour. Note that for \code{cumul = TRUE} the results do depend on
the order of qc_names. The flags on each row are cumulatively summed from
left to right. \code{additive} is used only if cumul = TRUE, otherwise
skipped.
}
\examples{
set.seed(5)
aa <- as.data.frame(replicate(
4, sample(c(0:2, NA), 20, replace = TRUE, prob = c(rep(0.3, 3), 0.1))))
names(aa) <- letters[1:4]

bb <- summary_QC(aa, letters[1:4])
is_add <- c(FALSE, TRUE, FALSE, TRUE)
summary_QC(aa, letters[1:4], na.as = c(NA, 0, NA, 2))
summary_QC(aa, letters[1:4], cumul = TRUE, additive = TRUE)
summary_QC(aa, letters[1:4], cumul = TRUE, additive = FALSE)
summary_QC(aa, letters[1:4], cumul = TRUE, additive = is_add)
cc <- summary_QC(aa, letters[1:4], cumul = FALSE, additive = is_add)
identical(bb, cc) # Argument additive is skipped when cumul = FALSE

library(ggplot2)
(xx <- summary_QC(aa, letters[1:4], cumul = TRUE, additive = TRUE,
plot = TRUE, flux = "CO2 flux"))
xx + ggplot2::theme(text = ggplot2::element_text(size = 20))

summary_QC(aa, rep(letters[1:4], 2),
cumul = TRUE, additive = TRUE, perc = FALSE)
summary_QC(aa, rep(letters[1:4], 2),
cumul = TRUE, additive = TRUE, plot = TRUE, perc = FALSE)
}
\seealso{
\code{\link{combn_QC}}.
}
